@@ -43,43 +48,29 @@
     },
     "homepage": "https://github.com/0xProject/protocol/tree/main/contracts/utils",
     "devDependencies": {
-        "@0x/abi-gen": "^5.8.5",
-        "@0x/contracts-gen": "^2.0.50",
-        "@0x/contracts-test-utils": "^5.4.59",
-        "@0x/dev-utils": "^5.0.2",
-        "@0x/order-utils": "^10.4.28",
-        "@0x/sol-compiler": "^4.8.2",
-        "@0x/types": "^3.3.7",
-        "@0x/web3-wrapper": "^8.0.1",
-        "@types/bn.js": "^4.11.0",
-        "@types/lodash": "4.14.104",
-        "@types/mocha": "^5.2.7",
+        "@typechain/ethers-v6": "^0.5.1",
+        "@typechain/hardhat": "^9.1.0",
+        "typechain": "^8.3.2",
+        "@types/lodash": "4.14.144",
+        "@types/mocha": "^9.1.0",
         "@types/node": "12.12.54",
-        "@typescript-eslint/eslint-plugin": "^5.38.0",
-        "@typescript-eslint/parser": "^5.38.0",
-        "chai": "^4.0.1",
-        "chai-as-promised": "^7.1.0",
-        "chai-bignumber": "^3.0.0",
-        "dirty-chai": "^2.0.1",
-        "eslint": "^8.23.1",
-        "eslint-config-prettier": "^8.5.0",
-        "ethereumjs-util": "^7.0.10",
+        "chai": "^4.5.0",
+        "chokidar": "^3.5.2",
+        "ethers": "^6.15.0",
+        "@nomicfoundation/hardhat-ethers": "^3.0.0",
+        "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
+        "hardhat": "^2.19.0",
         "lodash": "^4.17.11",
         "make-promises-safe": "^1.1.0",
-        "mocha": "^6.2.0",
+        "mocha": "^9.2.1",
         "npm-run-all": "^4.1.2",
-        "shx": "^0.2.2",
-        "typescript": "4.6.3"
+        "nyc": "^15.1.0",
+        "shx": "^0.4.0",
+        "ts-node": "^10.9.0",
+        "typescript": "^5.0.0"
     },
-    "dependencies": {
-        "@0x/base-contract": "^7.0.0",
-        "@0x/typescript-typings": "^5.3.1",
-        "@0x/utils": "^7.0.0",
-        "bn.js": "^4.11.8",
-        "ethereum-types": "^3.7.1"
-    },
+    "dependencies": {},
     "publishConfig": {
         "access": "public"
-    },
-    "gitHead": "4f91bfd907996b2f4dd383778b50c479c2602b56"
+    }
 }
