@@ -50,45 +66,31 @@
     },
     "homepage": "https://github.com/0xProject/protocol/tree/main/contracts/zero-ex",
     "devDependencies": {
-        "@0x/abi-gen": "^5.8.5",
-        "@0x/contract-addresses": "^8.12.0",
-        "@0x/contracts-erc20": "^3.3.57",
-        "@0x/contracts-gen": "^2.0.50",
-        "@0x/contracts-test-utils": "^5.4.59",
-        "@0x/dev-utils": "^5.0.2",
-        "@0x/order-utils": "^10.4.28",
-        "@0x/sol-compiler": "^4.8.2",
-        "@0x/ts-doc-gen": "^0.0.28",
-        "@typechain/ethers-v5": "^10.0.0",
-        "@types/isomorphic-fetch": "^0.0.35",
+        "@typechain/ethers-v6": "^0.5.1",
+        "@typechain/hardhat": "^9.1.0",
+        "typechain": "^8.3.2",
         "@types/lodash": "4.14.104",
         "@types/mocha": "^5.2.7",
-        "@types/prompts": "^2.0.9",
-        "@typescript-eslint/eslint-plugin": "^5.38.0",
-        "@typescript-eslint/parser": "^5.38.0",
-        "eslint": "^8.23.1",
-        "eslint-config-prettier": "^8.5.0",
-        "isomorphic-fetch": "^3.0.0",
+        "@types/node": "12.12.54",
+        "@nomicfoundation/hardhat-ethers": "^3.0.0",
+        "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
+        "hardhat": "^2.19.0",
+        "chai": "^4.5.0",
+        "dirty-chai": "^2.0.1",
+        "ethereumjs-util": "^7.0.10",
         "lodash": "^4.17.11",
-        "mocha": "^6.2.0",
+        "make-promises-safe": "^1.1.0",
+        "mocha": "^10.2.0",
         "npm-run-all": "^4.1.2",
-        "prompts": "^2.4.0",
-        "shx": "^0.2.2",
-        "typechain": "^8.0.0",
-        "typedoc": "~0.16.11",
-        "typescript": "4.6.3"
+        "nyc": "^15.1.0",
+        "shx": "^0.4.0",
+        "ts-node": "^10.9.0",
+        "typescript": "^5.0.0"
     },
     "dependencies": {
-        "@0x/base-contract": "^7.0.0",
-        "@0x/protocol-utils": "^11.24.1",
-        "@0x/subproviders": "^8.0.1",
-        "@0x/types": "^3.3.7",
-        "@0x/typescript-typings": "^5.3.1",
-        "@0x/utils": "^7.0.0",
-        "@0x/web3-wrapper": "^8.0.1",
         "ethereum-types": "^3.7.1",
         "ethereumjs-util": "^7.0.10",
-        "ethers": "~4.0.4"
+        "ethers": "^6.15.0"
     },
     "publishConfig": {
         "access": "public"
