@@ -59,29 +101,54 @@
         "ignoreDependencyVersionsForPackage": "contract-wrappers"
     },
     "devDependencies": {
-        "@0x-lerna-fork/lerna": "3.16.10",
         "@0xproject/npm-cli-login": "^0.0.11",
+        "@nomicfoundation/hardhat-chai-matchers": "^2.0.0",
+        "@nomicfoundation/hardhat-ethers": "^3.0.0",
+        "@types/lodash": "^4.14.200",
         "@types/minimatch": "^3.0.5",
+        "@typescript-eslint/eslint-plugin": "^8.38.0",
+        "@typescript-eslint/parser": "^8.38.0",
         "async-child-process": "^1.1.1",
         "check-md": "^1.0.2",
+        "copyfiles": "^2.4.1",
         "coveralls": "^3.0.0",
-        "ganache-cli": "6.12.2",
+        "eslint": "^9.32.0",
+        "eslint-config-prettier": "^10.1.8",
+        "eslint-plugin-prettier": "^5.5.3",
         "husky": "^8.0.3",
-        "lerna": "^3.0.0-beta.25",
         "lint-staged": "^13.1.0",
+        "make-promises-safe": "^1.1.0",
+        "mocha": "^10.2.0",
         "npm-run-all": "^4.1.2",
-        "prettier": "2.7.1",
-        "prettier-plugin-solidity": "^1.0.0",
+        "prettier": "^3.6.2",
+        "prettier-plugin-solidity": "1.4.3",
+        "shelljs": "^0.8.5",
+        "shx": "^0.3.4",
         "solhint": "^3.3.7",
-        "solhint-plugin-prettier": "^0.0.5",
+        "solhint-plugin-prettier": "0.1.0",
         "source-map-support": "^0.5.6",
-        "typescript": "4.6.3",
+        "ts-node": "^10.9.0",
+        "typescript": "^5.0.0",
         "wsrun": "^5.2.4"
     },
     "resolutions": {
-        "**/bignumber.js": "^9.0.2"
+        "**/bignumber.js": "^9.0.2",
+        "**/@0x/utils": "^7.0.0",
+        "**/ethers": "^6.15.0",
+        "**/@nomiclabs/hardhat-ethers": "npm:@nomicfoundation/hardhat-ethers@^3.0.0",
+        "**/shelljs": "^0.8.5",
+        "**/shx": "^0.3.4",
+        "**/typescript": "^5.0.0"
     },
     "dependencies": {
-        "@openzeppelin/contracts": "^4.8.1"
-    }
+        "@nomicfoundation/hardhat-ethers": "^3.0.0",
+        "@openzeppelin/contracts": "^4.8.1",
+        "@types/chai": "^4.3.0",
+        "@types/node": "^24.0.15",
+        "chai": "^4.5.0",
+        "ethers": "^6.15.0",
+        "hardhat": "^2.26.0",
+        "shx": "^0.3.4"
+    },
+    "version": "0.0.0"
 }
