@@ -4,21 +4,27 @@
     "engines": {
         "node": ">=6.12"
     },
-    "description": "0x smart contract compilation artifacts",
+    "description": "0x smart contract compilation artifacts in native Hardhat format for TypeChain",
     "main": "lib/src/index.js",
     "directories": {
         "test": "test"
     },
     "scripts": {
-        "artifacts_copy": "node lib/src/copy.js",
-        "artifacts_transform": "node lib/src/transform.js ./artifacts && prettier --write ./artifacts/*.json",
-        "artifacts_update": "yarn artifacts_copy && yarn artifacts_transform && yarn build",
-        "build": "yarn tsc -b",
+        "artifacts_copy": "node scripts/copy-artifacts.js",
+        "artifacts_update": "yarn artifacts_copy && yarn build",
+        "pre_build": "node scripts/copy-artifacts.js",
+        "build": "yarn pre_build && yarn tsc -b",
         "build:ci": "yarn build",
-        "test": "yarn run_mocha",
-        "test:ci": "yarn test",
-        "run_mocha": "mocha --require source-map-support/register --require make-promises-safe lib/test/**/*_test.js --bail --exit",
-        "clean": "shx rm -rf lib"
+        "test": "yarn pre_build && yarn test:hardhat",
+        "test:hardhat": "yarn pre_build && npx hardhat test",
+        "test:ci": "yarn pre_build && yarn test",
+        "test:coverage": "nyc --reporter=text --reporter=html --reporter=lcov yarn test",
+        "run_mocha": "npx mocha test/*.ts --require ts-node/register",
+        "clean": "shx rm -rf lib artifacts coverage .nyc_output *.log",
+        "clean:node-modules": "shx rm -rf node_modules",
+        "clean:all": "yarn clean && yarn clean:node-modules",
+        "lint": "eslint src test --ext .ts,.js",
+        "lint:fix": "eslint src test --ext .ts,.js --fix"
     },
     "repository": {
         "type": "git",
@@ -30,14 +36,14 @@
     },
     "homepage": "https://github.com/0xProject/protocol/tree/main/packages/contract-artifacts",
     "devDependencies": {
-        "@0x/utils": "^7.0.0",
-        "@types/minimatch": "^3.0.5",
         "@types/mocha": "^5.2.7",
-        "chai": "^4.0.1",
-        "lodash": "^4.17.11",
-        "mocha": "^6.2.0",
-        "shx": "^0.2.2",
-        "typescript": "4.6.3"
+        "chai": "^4.5.0",
+        "lodash": "^4.17.21",
+        "mocha": "^10.2.0",
+        "nyc": "^15.1.0",
+        "shx": "^0.4.0",
+        "ts-node": "^10.9.0",
+        "typescript": "^5.0.0"
     },
     "publishConfig": {
         "access": "public"
