@@ -44,24 +51,18 @@
     },
     "homepage": "https://github.com/0xProject/protocol/tree/main/packages/contract-wrappers",
     "devDependencies": {
-        "@0x/ts-doc-gen": "^0.0.28",
-        "@typescript-eslint/eslint-plugin": "^5.38.0",
-        "@typescript-eslint/parser": "^5.38.0",
-        "eslint": "^8.23.1",
-                "eslint-config-prettier": "^8.5.0",
+        "@typechain/ethers-v6": "^0.5.1",
+        "@typechain/hardhat": "^9.1.0",
+        "chai": "^4.5.0",
+        "hardhat": "^2.26.1",
+        "shx": "^0.4.0",
+        "typechain": "^8.3.2",
         "typedoc": "~0.16.11",
-        "typescript": "4.6.3"
+        "typescript": "^5.0.0"
     },
     "dependencies": {
-        "@0x/assert": "^3.0.36",
-        "@0x/base-contract": "^7.0.0",
-        "@0x/contract-addresses": "^8.12.0",
-        "@0x/json-schemas": "^6.4.4",
-        "@0x/types": "^3.3.7",
-        "@0x/utils": "^7.0.0",
-        "@0x/web3-wrapper": "^8.0.1",
         "ethereum-types": "^3.7.1",
-        "ethers": "~4.0.4"
+        "ethers": "^6.15.0"
     },
     "publishConfig": {
         "access": "public"
