@@ -9,14 +9,20 @@
     "types": "lib/src/index.d.ts",
     "scripts": {
         "build": "yarn tsc -b",
+        "build:ts": "yarn tsc -b",
         "build:ci": "yarn build",
         "watch": "tsc -w -p tsconfig.json",
         "publish:private": "yarn clean && yarn build && echo 'gitpkg publish disabled'",
-        "test": "yarn run_mocha",
+        "test": "yarn build && yarn test:hardhat",
+        "test:hardhat": "yarn build && npx hardhat test",
+        "test:ci": "yarn build && yarn test",
+        "test:coverage": "nyc --reporter=text --reporter=html --reporter=lcov yarn test",
         "rebuild_and_test": "run-s build test",
-        "run_mocha": "mocha --require source-map-support/register --require make-promises-safe lib/test/**/*_test.js --timeout 10000 --bail --exit",
-        "clean": "shx rm -rf lib generated_docs",
-        "lint": "eslint src",
+        "run_mocha": "npx mocha test/*.ts --require ts-node/register",
+        "clean": "shx rm -rf lib generated_docs coverage .nyc_output *.log",
+        "clean:node-modules": "shx rm -rf node_modules",
+        "clean:all": "yarn clean && yarn clean:node-modules",
+        "lint": "eslint src --ext .ts,.js",
         "fix": "eslint --fix --quiet src test",
         "diff_docs": "git diff --exit-code ./docs",
         "s3:sync_md_docs": "aws s3 sync ./docs s3://docs-markdown/${npm_package_name}/v${npm_package_version} --profile 0xproject --region us-east-1 --grants read=uri=http://acs.amazonaws.com/groups/global/AllUsers",
@@ -38,41 +44,25 @@
     },
     "homepage": "https://github.com/0xProject/protocol/tree/main/packages/protocol-utils",
     "devDependencies": {
-        "@0x/dev-utils": "^5.0.2",
-        "@0x/ts-doc-gen": "^0.0.28",
-        "@0x/types": "^3.3.7",
-        "@0x/typescript-typings": "^5.3.1",
-        "@types/bn.js": "^4.11.0",
-        "@types/lodash": "4.14.104",
-        "@types/mocha": "^5.2.7",
-        "@types/node": "12.12.54",
-        "@types/web3-provider-engine": "^14.0.0",
-        "@typescript-eslint/eslint-plugin": "^5.38.0",
-        "@typescript-eslint/parser": "^5.38.0",
-        "eslint": "^8.23.1",
-        "eslint-config-prettier": "^8.5.0",
-        "ethereum-types": "^3.7.1",
-        "mocha": "^6.2.0",
-        "npm-run-all": "^4.1.2",
-        "shx": "^0.2.2",
-        "sinon": "^4.0.0",
-        "typedoc": "~0.16.11",
-        "typescript": "4.6.3",
-        "web3-provider-engine": "14.0.6"
+        "@types/lodash": "^4.14.200",
+        "@types/mocha": "^9.1.0",
+        "@types/node": "^17.0.21",
+        "chai": "^4.5.0",
+        "chai-bignumber": "^3.0.0",
+        "dirty-chai": "^2.0.1",
+        "hardhat": "^2.26.0",
+        "lodash": "^4.17.21",
+        "mocha": "^10.2.0",
+        "npm-run-all": "^4.1.5",
+        "shx": "^0.3.2",
+        "ts-node": "^10.7.0",
+        "typescript": "^5.0.0"
     },
     "dependencies": {
-        "@0x/assert": "^3.0.36",
-        "@0x/contract-addresses": "^8.12.0",
-        "@0x/contract-wrappers": "^13.23.7",
-        "@0x/json-schemas": "^6.4.4",
-        "@0x/subproviders": "^8.0.1",
-        "@0x/utils": "^7.0.0",
-        "@0x/web3-wrapper": "^8.0.1",
-        "@typescript-eslint/parser": "^5.38.0",
-        "chai": "^4.0.1",
+        "chai": "^4.5.0",
         "ethereumjs-util": "^7.0.10",
-        "ethers": "~4.0.4",
-        "lodash": "^4.17.11"
+        "ethers": "^6.15.0",
+        "lodash": "^4.17.21"
     },
     "publishConfig": {
         "access": "public"
