@@ -3,13 +3,27 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-import { ContractArtifact } from 'ethereum-types';
+import type { TreasuryHardhatArtifacts, HardhatArtifact } from './types';
 
-import * as DefaultPoolOperator from '../generated-artifacts/DefaultPoolOperator.json';
-import * as ISablier from '../generated-artifacts/ISablier.json';
-import * as ZrxTreasury from '../generated-artifacts/ZrxTreasury.json';
-export const artifacts = {
-    ZrxTreasury: ZrxTreasury as ContractArtifact,
-    DefaultPoolOperator: DefaultPoolOperator as ContractArtifact,
-    ISablier: ISablier as ContractArtifact,
+// Import Hardhat artifacts
+import * as ZrxTreasury from '../artifacts/contracts/src/ZrxTreasury.sol/ZrxTreasury.json';
+import * as DefaultPoolOperator from '../artifacts/contracts/src/DefaultPoolOperator.sol/DefaultPoolOperator.json';
+import * as DummyERC20Token from '../artifacts/contracts/src/DummyERC20Token.sol/DummyERC20Token.json';
+import * as IStaking from '../artifacts/contracts/src/IStaking.sol/IStaking.json';
+import * as IZrxTreasury from '../artifacts/contracts/src/IZrxTreasury.sol/IZrxTreasury.json';
+import * as TreasuryStaking from '../artifacts/contracts/src/TreasuryStaking.sol/TreasuryStaking.json';
+import * as ISablier from '../artifacts/contracts/external/ISablier.sol/ISablier.json';
+
+/**
+ * Treasury 包的 Hardhat artifacts
+ * 使用 Hardhat 编译生成的 JSON 格式
+ */
+export const artifacts: TreasuryHardhatArtifacts = {
+    ZrxTreasury: ZrxTreasury as any as HardhatArtifact,
+    DefaultPoolOperator: DefaultPoolOperator as any as HardhatArtifact,
+    DummyERC20Token: DummyERC20Token as any as HardhatArtifact,
+    IStaking: IStaking as any as HardhatArtifact,
+    IZrxTreasury: IZrxTreasury as any as HardhatArtifact,
+    TreasuryStaking: TreasuryStaking as any as HardhatArtifact,
+    ISablier: ISablier as any as HardhatArtifact,
 };
