@@ -1,2 +1,10 @@
 export { artifacts } from './artifacts';
-export { DefaultPoolOperatorContract, ZrxTreasuryContract } from './wrappers';
+export {
+    IZrxTreasury as IZrxTreasuryContract,
+    IStaking as IStakingContract,
+    ZrxTreasury as ZrxTreasuryContract,
+    TreasuryStaking as TreasuryStakingContract,
+    DefaultPoolOperator as DefaultPoolOperatorContract,
+    DummyERC20Token as DummyERC20TokenContract,
+    ISablier as ISablierContract,
+} from './wrappers';
