@@ -3,6 +3,27 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-export * from '../generated-wrappers/default_pool_operator';
-export * from '../generated-wrappers/i_sablier';
-export * from '../generated-wrappers/zrx_treasury';
+// Export TypeChain generated contract types
+export {
+    IZrxTreasury,
+    IStaking,
+    ZrxTreasury,
+    TreasuryStaking,
+    DefaultPoolOperator,
+    DummyERC20Token,
+    ISablier,
+} from './typechain-types';
+
+// Export factory types for contract deployment
+export {
+    IZrxTreasury__factory,
+    IStaking__factory,
+    ZrxTreasury__factory,
+    TreasuryStaking__factory,
+    DefaultPoolOperator__factory,
+    DummyERC20Token__factory,
+    ISablier__factory,
+} from './typechain-types';
+
+// Re-export common types
+export * from './typechain-types/common';
