@@ -3,17 +3,35 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-import { ContractArtifact } from 'ethereum-types';
 
-import * as DefaultPoolOperator from '../test/generated-artifacts/DefaultPoolOperator.json';
-import * as ISablier from '../test/generated-artifacts/ISablier.json';
-import * as IStaking from '../test/generated-artifacts/IStaking.json';
-import * as IZrxTreasury from '../test/generated-artifacts/IZrxTreasury.json';
-import * as ZrxTreasury from '../test/generated-artifacts/ZrxTreasury.json';
+// Hardhat artifacts import - Use Hardhat format instead of Foundry
+import * as DummyERC20TokenHardhatArtifact from '../artifacts/contracts/src/DummyERC20Token.sol/DummyERC20Token.json';
+import * as DefaultPoolOperatorHardhatArtifact from '../artifacts/contracts/src/DefaultPoolOperator.sol/DefaultPoolOperator.json';
+import * as ZrxTreasuryHardhatArtifact from '../artifacts/contracts/src/ZrxTreasury.sol/ZrxTreasury.json';
+import * as TreasuryStakingHardhatArtifact from '../artifacts/contracts/src/TreasuryStaking.sol/TreasuryStaking.json';
+import * as IStakingHardhatArtifact from '../artifacts/contracts/src/IStaking.sol/IStaking.json';
+import * as IZrxTreasuryHardhatArtifact from '../artifacts/contracts/src/IZrxTreasury.sol/IZrxTreasury.json';
+
+// Export artifacts directly as Hardhat format (compatible with ethers.js)
 export const artifacts = {
-    ISablier: ISablier as ContractArtifact,
-    DefaultPoolOperator: DefaultPoolOperator as ContractArtifact,
-    IStaking: IStaking as ContractArtifact,
-    IZrxTreasury: IZrxTreasury as ContractArtifact,
-    ZrxTreasury: ZrxTreasury as ContractArtifact,
+    DummyERC20Token: DummyERC20TokenHardhatArtifact,
+    DefaultPoolOperator: DefaultPoolOperatorHardhatArtifact,
+    ZrxTreasury: ZrxTreasuryHardhatArtifact,
+    TreasuryStaking: TreasuryStakingHardhatArtifact,
+    IStaking: IStakingHardhatArtifact,
+    IZrxTreasury: IZrxTreasuryHardhatArtifact,
+};
+
+// Also export for compatibility
+export const erc20Artifacts = {
+    DummyERC20Token: DummyERC20TokenHardhatArtifact,
+};
+
+// Export treasury specific artifacts
+export const treasuryArtifacts = {
+    TreasuryStaking: TreasuryStakingHardhatArtifact,
+    ZrxTreasury: ZrxTreasuryHardhatArtifact,
+    DefaultPoolOperator: DefaultPoolOperatorHardhatArtifact,
+    IStaking: IStakingHardhatArtifact,
+    IZrxTreasury: IZrxTreasuryHardhatArtifact,
 };
