@@ -3,27 +3,83 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-import { ContractArtifact } from 'ethereum-types';
 
-import * as IOwnable from '../generated-artifacts/IOwnable.json';
-import * as LibBytes from '../generated-artifacts/LibBytes.json';
-import * as LibBytesRichErrors from '../generated-artifacts/LibBytesRichErrors.json';
-import * as LibOwnableRichErrors from '../generated-artifacts/LibOwnableRichErrors.json';
-import * as LibReentrancyGuardRichErrors from '../generated-artifacts/LibReentrancyGuardRichErrors.json';
-import * as LibRichErrors from '../generated-artifacts/LibRichErrors.json';
-import * as LibSafeMath from '../generated-artifacts/LibSafeMath.json';
-import * as LibSafeMathRichErrors from '../generated-artifacts/LibSafeMathRichErrors.json';
-import * as Ownable from '../generated-artifacts/Ownable.json';
-import * as ReentrancyGuard from '../generated-artifacts/ReentrancyGuard.json';
+// 导入 Hardhat 生成的 JSON artifacts
+import * as Authorizable from '../artifacts/contracts/src/Authorizable.sol/Authorizable.json';
+import * as IAuthorizable from '../artifacts/contracts/src/interfaces/IAuthorizable.sol/IAuthorizable.json';
+import * as IOwnable from '../artifacts/contracts/src/interfaces/IOwnable.sol/IOwnable.json';
+import * as LibAuthorizableRichErrors from '../artifacts/contracts/src/errors/LibAuthorizableRichErrors.sol/LibAuthorizableRichErrors.json';
+import * as LibBytesRichErrors from '../artifacts/contracts/src/errors/LibBytesRichErrors.sol/LibBytesRichErrors.json';
+import * as LibBytes from '../artifacts/contracts/src/LibBytes.sol/LibBytes.json';
+import * as LibMathRichErrors from '../artifacts/contracts/src/errors/LibMathRichErrors.sol/LibMathRichErrors.json';
+import * as LibMath from '../artifacts/contracts/src/LibMath.sol/LibMath.json';
+import * as LibOwnableRichErrors from '../artifacts/contracts/src/errors/LibOwnableRichErrors.sol/LibOwnableRichErrors.json';
+import * as LibReentrancyGuardRichErrors from '../artifacts/contracts/src/errors/LibReentrancyGuardRichErrors.sol/LibReentrancyGuardRichErrors.json';
+import * as LibRichErrors from '../artifacts/contracts/src/errors/LibRichErrors.sol/LibRichErrors.json';
+import * as LibSafeMathRichErrors from '../artifacts/contracts/src/errors/LibSafeMathRichErrors.sol/LibSafeMathRichErrors.json';
+import * as LibSafeMath from '../artifacts/contracts/src/LibSafeMath.sol/LibSafeMath.json';
+import * as Ownable from '../artifacts/contracts/src/Ownable.sol/Ownable.json';
+import * as ReentrancyGuard from '../artifacts/contracts/src/ReentrancyGuard.sol/ReentrancyGuard.json';
+
+// Import test contracts
+import * as TestLibBytes from '../artifacts/contracts/test/TestLibBytes.sol/TestLibBytes.json';
+import * as TestLibRichErrors from '../artifacts/contracts/test/TestLibRichErrors.sol/TestLibRichErrors.json';
+import * as TestLibSafeMath from '../artifacts/contracts/test/TestLibSafeMath.sol/TestLibSafeMath.json';
+import * as TestOwnable from '../artifacts/contracts/test/TestOwnable.sol/TestOwnable.json';
+import * as TestReentrancyGuard from '../artifacts/contracts/test/TestReentrancyGuard.sol/TestReentrancyGuard.json';
+
+/**
+ * Utils 包的 Hardhat artifacts
+ * 使用 Hardhat 编译生成的标准 artifacts 格式
+ */
 export const artifacts = {
-    IOwnable: IOwnable as ContractArtifact,
-    LibBytes: LibBytes as ContractArtifact,
-    LibBytesRichErrors: LibBytesRichErrors as ContractArtifact,
-    LibOwnableRichErrors: LibOwnableRichErrors as ContractArtifact,
-    LibReentrancyGuardRichErrors: LibReentrancyGuardRichErrors as ContractArtifact,
-    LibRichErrors: LibRichErrors as ContractArtifact,
-    LibSafeMath: LibSafeMath as ContractArtifact,
-    LibSafeMathRichErrors: LibSafeMathRichErrors as ContractArtifact,
-    Ownable: Ownable as ContractArtifact,
-    ReentrancyGuard: ReentrancyGuard as ContractArtifact,
+    // 主要合约
+    Authorizable,
+    IAuthorizable,
+    IOwnable,
+    LibBytes,
+    LibMath,
+    LibSafeMath,
+    Ownable,
+    ReentrancyGuard,
+
+    // 错误库
+    LibAuthorizableRichErrors,
+    LibBytesRichErrors,
+    LibMathRichErrors,
+    LibOwnableRichErrors,
+    LibReentrancyGuardRichErrors,
+    LibRichErrors,
+    LibSafeMathRichErrors,
+
+    // 测试合约
+    TestLibBytes,
+    TestLibRichErrors,
+    TestLibSafeMath,
+    TestOwnable,
+    TestReentrancyGuard,
+};
+
+// 为了向后兼容，也导出单独的 artifacts
+export {
+    Authorizable,
+    IAuthorizable,
+    IOwnable,
+    LibBytes,
+    LibMath,
+    LibSafeMath,
+    Ownable,
+    ReentrancyGuard,
+    LibAuthorizableRichErrors,
+    LibBytesRichErrors,
+    LibMathRichErrors,
+    LibOwnableRichErrors,
+    LibReentrancyGuardRichErrors,
+    LibRichErrors,
+    LibSafeMathRichErrors,
+    TestLibBytes,
+    TestLibRichErrors,
+    TestLibSafeMath,
+    TestOwnable,
+    TestReentrancyGuard,
 };
