@@ -1,13 +1,13 @@
-import { BigNumber, LibMathRevertErrors, SafeMathRevertErrors } from '@0x/utils';
+import { SafeMathRevertErrors } from '@0x/utils';
 
-const MAX_UINT256 = new BigNumber(2).pow(256).minus(1);
+const MAX_UINT256 = BigInt('0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff');
 
 /**
  * Add two `uint256` values. Reverts on overflow.
  */
-export function safeAdd(a: BigNumber, b: BigNumber): BigNumber {
-    const r = a.plus(b);
-    if (r.isGreaterThan(MAX_UINT256)) {
+export function safeAdd(a: bigint, b: bigint): bigint {
+    const r = a + b;
+    if (r > MAX_UINT256) {
         throw new SafeMathRevertErrors.Uint256BinOpError(SafeMathRevertErrors.BinOpErrorCodes.AdditionOverflow, a, b);
     }
     return r;
@@ -16,24 +16,23 @@ export function safeAdd(a: BigNumber, b: BigNumber): BigNumber {
 /**
  * Subract two `uint256` values. Reverts on overflow.
  */
-export function safeSub(a: BigNumber, b: BigNumber): BigNumber {
-    const r = a.minus(b);
-    if (r.isLessThan(0)) {
+export function safeSub(a: bigint, b: bigint): bigint {
+    if (a < b) {
         throw new SafeMathRevertErrors.Uint256BinOpError(
             SafeMathRevertErrors.BinOpErrorCodes.SubtractionUnderflow,
             a,
             b,
         );
     }
-    return r;
+    return a - b;
 }
 
 /**
  * Multiplies two `uint256` values. Reverts on overflow.
  */
-export function safeMul(a: BigNumber, b: BigNumber): BigNumber {
-    const r = a.times(b);
-    if (r.isGreaterThan(MAX_UINT256)) {
+export function safeMul(a: bigint, b: bigint): bigint {
+    const r = a * b;
+    if (r > MAX_UINT256) {
         throw new SafeMathRevertErrors.Uint256BinOpError(
             SafeMathRevertErrors.BinOpErrorCodes.MultiplicationOverflow,
             a,
@@ -46,9 +45,9 @@ export function safeMul(a: BigNumber, b: BigNumber): BigNumber {
 /**
  * Divides two `uint256` values. Reverts on division by zero.
  */
-export function safeDiv(a: BigNumber, b: BigNumber): BigNumber {
-    if (b.isEqualTo(0)) {
+export function safeDiv(a: bigint, b: bigint): bigint {
+    if (b === BigInt(0)) {
         throw new SafeMathRevertErrors.Uint256BinOpError(SafeMathRevertErrors.BinOpErrorCodes.DivisionByZero, a, b);
     }
-    return a.dividedToIntegerBy(b);
+    return a / b;
 }
