@@ -3,13 +3,53 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-export * from '../generated-wrappers/i_ownable';
-export * from '../generated-wrappers/lib_bytes';
-export * from '../generated-wrappers/lib_bytes_rich_errors';
-export * from '../generated-wrappers/lib_ownable_rich_errors';
-export * from '../generated-wrappers/lib_reentrancy_guard_rich_errors';
-export * from '../generated-wrappers/lib_rich_errors';
-export * from '../generated-wrappers/lib_safe_math';
-export * from '../generated-wrappers/lib_safe_math_rich_errors';
-export * from '../generated-wrappers/ownable';
-export * from '../generated-wrappers/reentrancy_guard';
+
+// Export Hardhat artifacts directly
+import * as AuthorizableArtifact from '../artifacts/contracts/src/Authorizable.sol/Authorizable.json';
+import * as IAuthorizableArtifact from '../artifacts/contracts/src/interfaces/IAuthorizable.sol/IAuthorizable.json';
+import * as IOwnableArtifact from '../artifacts/contracts/src/interfaces/IOwnable.sol/IOwnable.json';
+import * as OwnableArtifact from '../artifacts/contracts/src/Ownable.sol/Ownable.json';
+import * as TestLibBytesArtifact from '../artifacts/contracts/test/TestLibBytes.sol/TestLibBytes.json';
+import * as TestLibRichErrorsArtifact from '../artifacts/contracts/test/TestLibRichErrors.sol/TestLibRichErrors.json';
+import * as TestLibSafeMathArtifact from '../artifacts/contracts/test/TestLibSafeMath.sol/TestLibSafeMath.json';
+import * as TestOwnableArtifact from '../artifacts/contracts/test/TestOwnable.sol/TestOwnable.json';
+import * as TestReentrancyGuardArtifact from '../artifacts/contracts/test/TestReentrancyGuard.sol/TestReentrancyGuard.json';
+
+// Export library artifacts
+import * as LibBytesArtifact from '../artifacts/contracts/src/LibBytes.sol/LibBytes.json';
+import * as LibMathArtifact from '../artifacts/contracts/src/LibMath.sol/LibMath.json';
+import * as LibSafeMathArtifact from '../artifacts/contracts/src/LibSafeMath.sol/LibSafeMath.json';
+import * as ReentrancyGuardArtifact from '../artifacts/contracts/src/ReentrancyGuard.sol/ReentrancyGuard.json';
+
+// Export rich error artifacts
+import * as LibAuthorizableRichErrorsArtifact from '../artifacts/contracts/src/errors/LibAuthorizableRichErrors.sol/LibAuthorizableRichErrors.json';
+import * as LibBytesRichErrorsArtifact from '../artifacts/contracts/src/errors/LibBytesRichErrors.sol/LibBytesRichErrors.json';
+import * as LibMathRichErrorsArtifact from '../artifacts/contracts/src/errors/LibMathRichErrors.sol/LibMathRichErrors.json';
+import * as LibOwnableRichErrorsArtifact from '../artifacts/contracts/src/errors/LibOwnableRichErrors.sol/LibOwnableRichErrors.json';
+import * as LibReentrancyGuardRichErrorsArtifact from '../artifacts/contracts/src/errors/LibReentrancyGuardRichErrors.sol/LibReentrancyGuardRichErrors.json';
+import * as LibRichErrorsArtifact from '../artifacts/contracts/src/errors/LibRichErrors.sol/LibRichErrors.json';
+import * as LibSafeMathRichErrorsArtifact from '../artifacts/contracts/src/errors/LibSafeMathRichErrors.sol/LibSafeMathRichErrors.json';
+
+// Re-export all artifacts
+export {
+    AuthorizableArtifact,
+    IAuthorizableArtifact,
+    IOwnableArtifact,
+    OwnableArtifact,
+    TestLibBytesArtifact,
+    TestLibRichErrorsArtifact,
+    TestLibSafeMathArtifact,
+    TestOwnableArtifact,
+    TestReentrancyGuardArtifact,
+    LibBytesArtifact,
+    LibMathArtifact,
+    LibSafeMathArtifact,
+    ReentrancyGuardArtifact,
+    LibAuthorizableRichErrorsArtifact,
+    LibBytesRichErrorsArtifact,
+    LibMathRichErrorsArtifact,
+    LibOwnableRichErrorsArtifact,
+    LibReentrancyGuardRichErrorsArtifact,
+    LibRichErrorsArtifact,
+    LibSafeMathRichErrorsArtifact,
+};
