@@ -3,93 +3,83 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-import { ContractArtifact } from 'ethereum-types';
 
-import * as AuthorizableV06 from '../test/generated-artifacts/AuthorizableV06.json';
-import * as AuthorizableV08 from '../test/generated-artifacts/AuthorizableV08.json';
-import * as IAuthorizableV06 from '../test/generated-artifacts/IAuthorizableV06.json';
-import * as IAuthorizableV08 from '../test/generated-artifacts/IAuthorizableV08.json';
-import * as IOwnable from '../test/generated-artifacts/IOwnable.json';
-import * as IOwnableV06 from '../test/generated-artifacts/IOwnableV06.json';
-import * as IOwnableV08 from '../test/generated-artifacts/IOwnableV08.json';
-import * as LibAuthorizableRichErrorsV06 from '../test/generated-artifacts/LibAuthorizableRichErrorsV06.json';
-import * as LibAuthorizableRichErrorsV08 from '../test/generated-artifacts/LibAuthorizableRichErrorsV08.json';
-import * as LibBytes from '../test/generated-artifacts/LibBytes.json';
-import * as LibBytesRichErrors from '../test/generated-artifacts/LibBytesRichErrors.json';
-import * as LibBytesRichErrorsV06 from '../test/generated-artifacts/LibBytesRichErrorsV06.json';
-import * as LibBytesRichErrorsV08 from '../test/generated-artifacts/LibBytesRichErrorsV08.json';
-import * as LibBytesV06 from '../test/generated-artifacts/LibBytesV06.json';
-import * as LibBytesV08 from '../test/generated-artifacts/LibBytesV08.json';
-import * as LibMathRichErrorsV06 from '../test/generated-artifacts/LibMathRichErrorsV06.json';
-import * as LibMathRichErrorsV08 from '../test/generated-artifacts/LibMathRichErrorsV08.json';
-import * as LibMathV06 from '../test/generated-artifacts/LibMathV06.json';
-import * as LibMathV08 from '../test/generated-artifacts/LibMathV08.json';
-import * as LibOwnableRichErrors from '../test/generated-artifacts/LibOwnableRichErrors.json';
-import * as LibOwnableRichErrorsV06 from '../test/generated-artifacts/LibOwnableRichErrorsV06.json';
-import * as LibOwnableRichErrorsV08 from '../test/generated-artifacts/LibOwnableRichErrorsV08.json';
-import * as LibReentrancyGuardRichErrors from '../test/generated-artifacts/LibReentrancyGuardRichErrors.json';
-import * as LibReentrancyGuardRichErrorsV06 from '../test/generated-artifacts/LibReentrancyGuardRichErrorsV06.json';
-import * as LibReentrancyGuardRichErrorsV08 from '../test/generated-artifacts/LibReentrancyGuardRichErrorsV08.json';
-import * as LibRichErrors from '../test/generated-artifacts/LibRichErrors.json';
-import * as LibRichErrorsV06 from '../test/generated-artifacts/LibRichErrorsV06.json';
-import * as LibRichErrorsV08 from '../test/generated-artifacts/LibRichErrorsV08.json';
-import * as LibSafeMath from '../test/generated-artifacts/LibSafeMath.json';
-import * as LibSafeMathRichErrors from '../test/generated-artifacts/LibSafeMathRichErrors.json';
-import * as LibSafeMathRichErrorsV06 from '../test/generated-artifacts/LibSafeMathRichErrorsV06.json';
-import * as LibSafeMathRichErrorsV08 from '../test/generated-artifacts/LibSafeMathRichErrorsV08.json';
-import * as LibSafeMathV06 from '../test/generated-artifacts/LibSafeMathV06.json';
-import * as Ownable from '../test/generated-artifacts/Ownable.json';
-import * as OwnableV06 from '../test/generated-artifacts/OwnableV06.json';
-import * as OwnableV08 from '../test/generated-artifacts/OwnableV08.json';
-import * as ReentrancyGuard from '../test/generated-artifacts/ReentrancyGuard.json';
-import * as ReentrancyGuardV08 from '../test/generated-artifacts/ReentrancyGuardV08.json';
-import * as TestLibBytes from '../test/generated-artifacts/TestLibBytes.json';
-import * as TestLibRichErrors from '../test/generated-artifacts/TestLibRichErrors.json';
-import * as TestLibSafeMath from '../test/generated-artifacts/TestLibSafeMath.json';
-import * as TestOwnable from '../test/generated-artifacts/TestOwnable.json';
-import * as TestReentrancyGuard from '../test/generated-artifacts/TestReentrancyGuard.json';
+// 导入 Hardhat 生成的 JSON artifacts
+import * as Authorizable from '../artifacts/contracts/src/Authorizable.sol/Authorizable.json';
+import * as IAuthorizable from '../artifacts/contracts/src/interfaces/IAuthorizable.sol/IAuthorizable.json';
+import * as IOwnable from '../artifacts/contracts/src/interfaces/IOwnable.sol/IOwnable.json';
+import * as LibAuthorizableRichErrors from '../artifacts/contracts/src/errors/LibAuthorizableRichErrors.sol/LibAuthorizableRichErrors.json';
+import * as LibBytesRichErrors from '../artifacts/contracts/src/errors/LibBytesRichErrors.sol/LibBytesRichErrors.json';
+import * as LibBytes from '../artifacts/contracts/src/LibBytes.sol/LibBytes.json';
+import * as LibMathRichErrors from '../artifacts/contracts/src/errors/LibMathRichErrors.sol/LibMathRichErrors.json';
+import * as LibMath from '../artifacts/contracts/src/LibMath.sol/LibMath.json';
+import * as LibOwnableRichErrors from '../artifacts/contracts/src/errors/LibOwnableRichErrors.sol/LibOwnableRichErrors.json';
+import * as LibReentrancyGuardRichErrors from '../artifacts/contracts/src/errors/LibReentrancyGuardRichErrors.sol/LibReentrancyGuardRichErrors.json';
+import * as LibRichErrors from '../artifacts/contracts/src/errors/LibRichErrors.sol/LibRichErrors.json';
+import * as LibSafeMathRichErrors from '../artifacts/contracts/src/errors/LibSafeMathRichErrors.sol/LibSafeMathRichErrors.json';
+import * as LibSafeMath from '../artifacts/contracts/src/LibSafeMath.sol/LibSafeMath.json';
+import * as Ownable from '../artifacts/contracts/src/Ownable.sol/Ownable.json';
+import * as ReentrancyGuard from '../artifacts/contracts/src/ReentrancyGuard.sol/ReentrancyGuard.json';
+
+// Import test contracts
+import * as TestLibBytes from '../artifacts/contracts/test/TestLibBytes.sol/TestLibBytes.json';
+import * as TestLibRichErrors from '../artifacts/contracts/test/TestLibRichErrors.sol/TestLibRichErrors.json';
+import * as TestLibSafeMath from '../artifacts/contracts/test/TestLibSafeMath.sol/TestLibSafeMath.json';
+import * as TestOwnable from '../artifacts/contracts/test/TestOwnable.sol/TestOwnable.json';
+import * as TestReentrancyGuard from '../artifacts/contracts/test/TestReentrancyGuard.sol/TestReentrancyGuard.json';
+
+/**
+ * Utils 包的 Hardhat artifacts
+ * 使用 Hardhat 编译生成的标准 artifacts 格式
+ */
 export const artifacts = {
-    LibBytes: LibBytes as ContractArtifact,
-    LibBytesRichErrors: LibBytesRichErrors as ContractArtifact,
-    LibOwnableRichErrors: LibOwnableRichErrors as ContractArtifact,
-    LibReentrancyGuardRichErrors: LibReentrancyGuardRichErrors as ContractArtifact,
-    LibRichErrors: LibRichErrors as ContractArtifact,
-    LibSafeMath: LibSafeMath as ContractArtifact,
-    LibSafeMathRichErrors: LibSafeMathRichErrors as ContractArtifact,
-    Ownable: Ownable as ContractArtifact,
-    ReentrancyGuard: ReentrancyGuard as ContractArtifact,
-    IOwnable: IOwnable as ContractArtifact,
-    AuthorizableV06: AuthorizableV06 as ContractArtifact,
-    LibBytesV06: LibBytesV06 as ContractArtifact,
-    LibMathV06: LibMathV06 as ContractArtifact,
-    LibSafeMathV06: LibSafeMathV06 as ContractArtifact,
-    OwnableV06: OwnableV06 as ContractArtifact,
-    LibAuthorizableRichErrorsV06: LibAuthorizableRichErrorsV06 as ContractArtifact,
-    LibBytesRichErrorsV06: LibBytesRichErrorsV06 as ContractArtifact,
-    LibMathRichErrorsV06: LibMathRichErrorsV06 as ContractArtifact,
-    LibOwnableRichErrorsV06: LibOwnableRichErrorsV06 as ContractArtifact,
-    LibReentrancyGuardRichErrorsV06: LibReentrancyGuardRichErrorsV06 as ContractArtifact,
-    LibRichErrorsV06: LibRichErrorsV06 as ContractArtifact,
-    LibSafeMathRichErrorsV06: LibSafeMathRichErrorsV06 as ContractArtifact,
-    IAuthorizableV06: IAuthorizableV06 as ContractArtifact,
-    IOwnableV06: IOwnableV06 as ContractArtifact,
-    AuthorizableV08: AuthorizableV08 as ContractArtifact,
-    LibBytesV08: LibBytesV08 as ContractArtifact,
-    LibMathV08: LibMathV08 as ContractArtifact,
-    OwnableV08: OwnableV08 as ContractArtifact,
-    ReentrancyGuardV08: ReentrancyGuardV08 as ContractArtifact,
-    LibAuthorizableRichErrorsV08: LibAuthorizableRichErrorsV08 as ContractArtifact,
-    LibBytesRichErrorsV08: LibBytesRichErrorsV08 as ContractArtifact,
-    LibMathRichErrorsV08: LibMathRichErrorsV08 as ContractArtifact,
-    LibOwnableRichErrorsV08: LibOwnableRichErrorsV08 as ContractArtifact,
-    LibReentrancyGuardRichErrorsV08: LibReentrancyGuardRichErrorsV08 as ContractArtifact,
-    LibRichErrorsV08: LibRichErrorsV08 as ContractArtifact,
-    LibSafeMathRichErrorsV08: LibSafeMathRichErrorsV08 as ContractArtifact,
-    IAuthorizableV08: IAuthorizableV08 as ContractArtifact,
-    IOwnableV08: IOwnableV08 as ContractArtifact,
-    TestLibBytes: TestLibBytes as ContractArtifact,
-    TestLibRichErrors: TestLibRichErrors as ContractArtifact,
-    TestLibSafeMath: TestLibSafeMath as ContractArtifact,
-    TestOwnable: TestOwnable as ContractArtifact,
-    TestReentrancyGuard: TestReentrancyGuard as ContractArtifact,
+    // 主要合约
+    Authorizable,
+    IAuthorizable,
+    IOwnable,
+    LibBytes,
+    LibMath,
+    LibSafeMath,
+    Ownable,
+    ReentrancyGuard,
+
+    // 错误库
+    LibAuthorizableRichErrors,
+    LibBytesRichErrors,
+    LibMathRichErrors,
+    LibOwnableRichErrors,
+    LibReentrancyGuardRichErrors,
+    LibRichErrors,
+    LibSafeMathRichErrors,
+
+    // 测试合约
+    TestLibBytes,
+    TestLibRichErrors,
+    TestLibSafeMath,
+    TestOwnable,
+    TestReentrancyGuard,
+};
+
+// 为了向后兼容，也导出单独的 artifacts
+export {
+    Authorizable,
+    IAuthorizable,
+    IOwnable,
+    LibBytes,
+    LibMath,
+    LibSafeMath,
+    Ownable,
+    ReentrancyGuard,
+    LibAuthorizableRichErrors,
+    LibBytesRichErrors,
+    LibMathRichErrors,
+    LibOwnableRichErrors,
+    LibReentrancyGuardRichErrors,
+    LibRichErrors,
+    LibSafeMathRichErrors,
+    TestLibBytes,
+    TestLibRichErrors,
+    TestLibSafeMath,
+    TestOwnable,
+    TestReentrancyGuard,
 };
