@@ -1,26 +1,48 @@
-import { constants, describe, expect } from '@0x/contracts-test-utils';
+import { expect } from 'chai';
+import { ethers } from 'hardhat';
 import { SafeMathRevertErrors } from '@0x/utils';
 
 import { safeAdd, safeDiv, safeMul, safeSub } from '../src/reference_functions';
 
+// 现代化常量定义
+const MAX_UINT256 = ethers.MaxUint256;
+const ZERO_AMOUNT = 0n;
+const ONE_ETHER = ethers.parseEther('1');
+
+// Helper function for bigint square root calculation
+function bigintSqrt(value: bigint): bigint {
+    if (value === BigInt(0)) return BigInt(0);
+    if (value < BigInt(4)) return BigInt(1);
+
+    let x = value;
+    let y = (value + BigInt(1)) / BigInt(2);
+
+    while (y < x) {
+        x = y;
+        y = (y + value / y) / BigInt(2);
+    }
+
+    return x;
+}
+
 describe('Reference Functions', () => {
-    const { ONE_ETHER, MAX_UINT256, ZERO_AMOUNT } = constants;
     const DEFAULT_VALUES = {
-        a: ONE_ETHER.times(2),
+        a: ONE_ETHER * BigInt(2),
         b: ONE_ETHER,
     };
     describe('SafeMath', () => {
         describe('safeAdd', () => {
             it('adds two numbers', () => {
                 const { a, b } = DEFAULT_VALUES;
-                const expected = a.plus(b);
+                const expected = a + b;
                 const actual = safeAdd(a, b);
-                expect(actual).to.bignumber.eq(expected);
+                // Convert both to strings for comparison to avoid BigNumber comparison issues
+                expect(actual.toString()).to.equal(expected.toString());
             });
 
             it('reverts on overflow', () => {
-                const a = MAX_UINT256.dividedToIntegerBy(2);
-                const b = MAX_UINT256.dividedToIntegerBy(2).plus(2);
+                const a = MAX_UINT256 / BigInt(2);
+                const b = MAX_UINT256 / BigInt(2) + BigInt(2);
                 const expectedError = new SafeMathRevertErrors.Uint256BinOpError(
                     SafeMathRevertErrors.BinOpErrorCodes.AdditionOverflow,
                     a,
@@ -33,34 +55,35 @@ describe('Reference Functions', () => {
         describe('safeSub', () => {
             it('subracts two numbers', () => {
                 const { a, b } = DEFAULT_VALUES;
-                const expected = a.minus(b);
+                const expected = a - b;
                 const actual = safeSub(a, b);
-                expect(actual).to.bignumber.eq(expected);
+                // Convert both to strings for comparison
+                expect(actual.toString()).to.equal(expected.toString());
             });
 
             it('reverts on underflow', () => {
-                const a = MAX_UINT256.dividedToIntegerBy(2);
-                const b = MAX_UINT256.dividedToIntegerBy(2).plus(2);
+                const { a, b } = DEFAULT_VALUES;
                 const expectedError = new SafeMathRevertErrors.Uint256BinOpError(
                     SafeMathRevertErrors.BinOpErrorCodes.SubtractionUnderflow,
-                    a,
                     b,
+                    a,
                 );
-                expect(() => safeSub(a, b)).to.throw(expectedError.message);
+                expect(() => safeSub(b, a)).to.throw(expectedError.message);
             });
         });
 
         describe('safeMul', () => {
             it('multiplies two numbers', () => {
                 const { a, b } = DEFAULT_VALUES;
-                const expected = a.times(b);
+                const expected = a * b;
                 const actual = safeMul(a, b);
-                expect(actual).to.bignumber.eq(expected);
+                // Convert both to strings for comparison
+                expect(actual.toString()).to.equal(expected.toString());
             });
 
             it('reverts on overflow', () => {
-                const a = MAX_UINT256.dividedToIntegerBy(2);
-                const b = MAX_UINT256.dividedToIntegerBy(2).plus(2);
+                const a = bigintSqrt(MAX_UINT256) + BigInt(1);
+                const b = a;
                 const expectedError = new SafeMathRevertErrors.Uint256BinOpError(
                     SafeMathRevertErrors.BinOpErrorCodes.MultiplicationOverflow,
                     a,
@@ -73,20 +96,20 @@ describe('Reference Functions', () => {
         describe('safeDiv', () => {
             it('multiplies two numbers', () => {
                 const { a, b } = DEFAULT_VALUES;
-                const expected = a.times(b);
-                const actual = safeMul(a, b);
-                expect(actual).to.bignumber.eq(expected);
+                const expected = a / b;
+                const actual = safeDiv(a, b);
+                // Convert both to strings for comparison
+                expect(actual.toString()).to.equal(expected.toString());
             });
 
             it('reverts if denominator is zero', () => {
-                const a = MAX_UINT256.dividedToIntegerBy(2);
-                const b = ZERO_AMOUNT;
+                const { a } = DEFAULT_VALUES;
                 const expectedError = new SafeMathRevertErrors.Uint256BinOpError(
                     SafeMathRevertErrors.BinOpErrorCodes.DivisionByZero,
                     a,
-                    b,
+                    ZERO_AMOUNT,
                 );
-                expect(() => safeDiv(a, b)).to.throw(expectedError.message);
+                expect(() => safeDiv(a, ZERO_AMOUNT)).to.throw(expectedError.message);
             });
         });
     });
