@@ -3,46 +3,27 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-export * from '../test/generated-wrappers/authorizable_v06';
-export * from '../test/generated-wrappers/authorizable_v08';
-export * from '../test/generated-wrappers/i_authorizable_v06';
-export * from '../test/generated-wrappers/i_authorizable_v08';
-export * from '../test/generated-wrappers/i_ownable';
-export * from '../test/generated-wrappers/i_ownable_v06';
-export * from '../test/generated-wrappers/i_ownable_v08';
-export * from '../test/generated-wrappers/lib_authorizable_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_authorizable_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_bytes';
-export * from '../test/generated-wrappers/lib_bytes_rich_errors';
-export * from '../test/generated-wrappers/lib_bytes_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_bytes_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_bytes_v06';
-export * from '../test/generated-wrappers/lib_bytes_v08';
-export * from '../test/generated-wrappers/lib_math_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_math_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_math_v06';
-export * from '../test/generated-wrappers/lib_math_v08';
-export * from '../test/generated-wrappers/lib_ownable_rich_errors';
-export * from '../test/generated-wrappers/lib_ownable_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_ownable_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_reentrancy_guard_rich_errors';
-export * from '../test/generated-wrappers/lib_reentrancy_guard_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_reentrancy_guard_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_rich_errors';
-export * from '../test/generated-wrappers/lib_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_safe_math';
-export * from '../test/generated-wrappers/lib_safe_math_rich_errors';
-export * from '../test/generated-wrappers/lib_safe_math_rich_errors_v06';
-export * from '../test/generated-wrappers/lib_safe_math_rich_errors_v08';
-export * from '../test/generated-wrappers/lib_safe_math_v06';
-export * from '../test/generated-wrappers/ownable';
-export * from '../test/generated-wrappers/ownable_v06';
-export * from '../test/generated-wrappers/ownable_v08';
-export * from '../test/generated-wrappers/reentrancy_guard';
-export * from '../test/generated-wrappers/reentrancy_guard_v08';
-export * from '../test/generated-wrappers/test_lib_bytes';
-export * from '../test/generated-wrappers/test_lib_rich_errors';
-export * from '../test/generated-wrappers/test_lib_safe_math';
-export * from '../test/generated-wrappers/test_ownable';
-export * from '../test/generated-wrappers/test_reentrancy_guard';
+// Export TypeChain generated contract types
+export { Authorizable } from '../src/typechain-types/src/Authorizable';
+export { IAuthorizable } from '../src/typechain-types/src/interfaces/IAuthorizable';
+export { IOwnable } from '../src/typechain-types/src/interfaces/IOwnable';
+export { Ownable } from '../src/typechain-types/src/Ownable';
+export { TestLibBytes } from '../src/typechain-types/test/TestLibBytes';
+export { TestLibRichErrors } from '../src/typechain-types/test/TestLibRichErrors';
+export { TestLibSafeMath } from '../src/typechain-types/test/TestLibSafeMath';
+export { TestOwnable } from '../src/typechain-types/test/TestOwnable';
+export { TestReentrancyGuard } from '../src/typechain-types/test/TestReentrancyGuard';
+
+// Export factory types for contract deployment
+export { Authorizable__factory } from '../src/typechain-types/factories/src/Authorizable__factory';
+export { IAuthorizable__factory } from '../src/typechain-types/factories/src/interfaces/IAuthorizable__factory';
+export { IOwnable__factory } from '../src/typechain-types/factories/src/interfaces/IOwnable__factory';
+export { Ownable__factory } from '../src/typechain-types/factories/src/Ownable__factory';
+export { TestLibBytes__factory } from '../src/typechain-types/factories/test/TestLibBytes__factory';
+export { TestLibRichErrors__factory } from '../src/typechain-types/factories/test/TestLibRichErrors__factory';
+export { TestLibSafeMath__factory } from '../src/typechain-types/factories/test/TestLibSafeMath__factory';
+export { TestOwnable__factory } from '../src/typechain-types/factories/test/TestOwnable__factory';
+export { TestReentrancyGuard__factory } from '../src/typechain-types/factories/test/TestReentrancyGuard__factory';
+
+// Re-export common types
+export * from '../src/typechain-types/common';
