@@ -3,91 +3,104 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-import { ContractArtifact } from 'ethereum-types';
 
-import * as AffiliateFeeTransformer from '../generated-artifacts/AffiliateFeeTransformer.json';
-import * as AvalancheBridgeAdapter from '../generated-artifacts/AvalancheBridgeAdapter.json';
-import * as BaseBridgeAdapter from '../generated-artifacts/BaseBridgeAdapter.json';
-import * as BaseGoerliBridgeAdapter from '../generated-artifacts/BaseGoerliBridgeAdapter.json';
-import * as BatchFillNativeOrdersFeature from '../generated-artifacts/BatchFillNativeOrdersFeature.json';
-import * as BSCBridgeAdapter from '../generated-artifacts/BSCBridgeAdapter.json';
-import * as CeloBridgeAdapter from '../generated-artifacts/CeloBridgeAdapter.json';
-import * as CurveLiquidityProvider from '../generated-artifacts/CurveLiquidityProvider.json';
-import * as EthereumBridgeAdapter from '../generated-artifacts/EthereumBridgeAdapter.json';
-import * as FantomBridgeAdapter from '../generated-artifacts/FantomBridgeAdapter.json';
-import * as FeeCollector from '../generated-artifacts/FeeCollector.json';
-import * as FeeCollectorController from '../generated-artifacts/FeeCollectorController.json';
-import * as FillQuoteTransformer from '../generated-artifacts/FillQuoteTransformer.json';
-import * as FullMigration from '../generated-artifacts/FullMigration.json';
-import * as IBatchFillNativeOrdersFeature from '../generated-artifacts/IBatchFillNativeOrdersFeature.json';
-import * as IERC20Transformer from '../generated-artifacts/IERC20Transformer.json';
-import * as IFlashWallet from '../generated-artifacts/IFlashWallet.json';
-import * as ILiquidityProviderFeature from '../generated-artifacts/ILiquidityProviderFeature.json';
-import * as IMultiplexFeature from '../generated-artifacts/IMultiplexFeature.json';
-import * as INativeOrdersFeature from '../generated-artifacts/INativeOrdersFeature.json';
-import * as InitialMigration from '../generated-artifacts/InitialMigration.json';
-import * as IOtcOrdersFeature from '../generated-artifacts/IOtcOrdersFeature.json';
-import * as IOwnableFeature from '../generated-artifacts/IOwnableFeature.json';
-import * as ISimpleFunctionRegistryFeature from '../generated-artifacts/ISimpleFunctionRegistryFeature.json';
-import * as ITransformERC20Feature from '../generated-artifacts/ITransformERC20Feature.json';
-import * as IZeroEx from '../generated-artifacts/IZeroEx.json';
-import * as LiquidityProviderFeature from '../generated-artifacts/LiquidityProviderFeature.json';
-import * as LogMetadataTransformer from '../generated-artifacts/LogMetadataTransformer.json';
-import * as MetaTransactionsFeature from '../generated-artifacts/MetaTransactionsFeature.json';
-import * as MetaTransactionsFeatureV2 from '../generated-artifacts/MetaTransactionsFeatureV2.json';
-import * as MultiplexFeature from '../generated-artifacts/MultiplexFeature.json';
-import * as NativeOrdersFeature from '../generated-artifacts/NativeOrdersFeature.json';
-import * as OptimismBridgeAdapter from '../generated-artifacts/OptimismBridgeAdapter.json';
-import * as OtcOrdersFeature from '../generated-artifacts/OtcOrdersFeature.json';
-import * as OwnableFeature from '../generated-artifacts/OwnableFeature.json';
-import * as PayTakerTransformer from '../generated-artifacts/PayTakerTransformer.json';
-import * as PolygonBridgeAdapter from '../generated-artifacts/PolygonBridgeAdapter.json';
-import * as PositiveSlippageFeeTransformer from '../generated-artifacts/PositiveSlippageFeeTransformer.json';
-import * as SimpleFunctionRegistryFeature from '../generated-artifacts/SimpleFunctionRegistryFeature.json';
-import * as TransformERC20Feature from '../generated-artifacts/TransformERC20Feature.json';
-import * as WethTransformer from '../generated-artifacts/WethTransformer.json';
-import * as ZeroEx from '../generated-artifacts/ZeroEx.json';
+// 核心合约
+import * as ZeroEx from '../artifacts/contracts/src/ZeroEx.sol/ZeroEx.json';
+import * as ZeroExOptimized from '../artifacts/contracts/src/ZeroExOptimized.sol/ZeroExOptimized.json';
+import * as IZeroEx from '../artifacts/contracts/src/IZeroEx.sol/IZeroEx.json';
+import * as FullMigration from '../artifacts/contracts/src/migrations/FullMigration.sol/FullMigration.json';
+import * as InitialMigration from '../artifacts/contracts/src/migrations/InitialMigration.sol/InitialMigration.json';
+import * as FlashWallet from '../artifacts/contracts/src/external/FlashWallet.sol/FlashWallet.json';
+import * as IFlashWallet from '../artifacts/contracts/src/external/IFlashWallet.sol/IFlashWallet.json';
+
+// Features
+import * as BatchFillNativeOrdersFeature from '../artifacts/contracts/src/features/BatchFillNativeOrdersFeature.sol/BatchFillNativeOrdersFeature.json';
+import * as IBatchFillNativeOrdersFeature from '../artifacts/contracts/src/features/interfaces/IBatchFillNativeOrdersFeature.sol/IBatchFillNativeOrdersFeature.json';
+import * as LiquidityProviderFeature from '../artifacts/contracts/src/features/LiquidityProviderFeature.sol/LiquidityProviderFeature.json';
+import * as ILiquidityProviderFeature from '../artifacts/contracts/src/features/interfaces/ILiquidityProviderFeature.sol/ILiquidityProviderFeature.json';
+import * as MultiplexFeature from '../artifacts/contracts/src/features/multiplex/MultiplexFeature.sol/MultiplexFeature.json';
+import * as IMultiplexFeature from '../artifacts/contracts/src/features/interfaces/IMultiplexFeature.sol/IMultiplexFeature.json';
+import * as NativeOrdersFeature from '../artifacts/contracts/src/features/NativeOrdersFeature.sol/NativeOrdersFeature.json';
+import * as INativeOrdersFeature from '../artifacts/contracts/src/features/interfaces/INativeOrdersFeature.sol/INativeOrdersFeature.json';
+import * as OtcOrdersFeature from '../artifacts/contracts/src/features/OtcOrdersFeature.sol/OtcOrdersFeature.json';
+import * as IOtcOrdersFeature from '../artifacts/contracts/src/features/interfaces/IOtcOrdersFeature.sol/IOtcOrdersFeature.json';
+import * as OwnableFeature from '../artifacts/contracts/src/features/OwnableFeature.sol/OwnableFeature.json';
+import * as IOwnableFeature from '../artifacts/contracts/src/features/interfaces/IOwnableFeature.sol/IOwnableFeature.json';
+import * as SimpleFunctionRegistryFeature from '../artifacts/contracts/src/features/SimpleFunctionRegistryFeature.sol/SimpleFunctionRegistryFeature.json';
+import * as ISimpleFunctionRegistryFeature from '../artifacts/contracts/src/features/interfaces/ISimpleFunctionRegistryFeature.sol/ISimpleFunctionRegistryFeature.json';
+import * as TransformERC20Feature from '../artifacts/contracts/src/features/TransformERC20Feature.sol/TransformERC20Feature.json';
+import * as ITransformERC20Feature from '../artifacts/contracts/src/features/interfaces/ITransformERC20Feature.sol/ITransformERC20Feature.json';
+import * as MetaTransactionsFeature from '../artifacts/contracts/src/features/MetaTransactionsFeature.sol/MetaTransactionsFeature.json';
+import * as MetaTransactionsFeatureV2 from '../artifacts/contracts/src/features/MetaTransactionsFeatureV2.sol/MetaTransactionsFeatureV2.json';
+
+// Transformers
+import * as AffiliateFeeTransformer from '../artifacts/contracts/src/transformers/AffiliateFeeTransformer.sol/AffiliateFeeTransformer.json';
+import * as FillQuoteTransformer from '../artifacts/contracts/src/transformers/FillQuoteTransformer.sol/FillQuoteTransformer.json';
+import * as PayTakerTransformer from '../artifacts/contracts/src/transformers/PayTakerTransformer.sol/PayTakerTransformer.json';
+import * as PositiveSlippageFeeTransformer from '../artifacts/contracts/src/transformers/PositiveSlippageFeeTransformer.sol/PositiveSlippageFeeTransformer.json';
+import * as WethTransformer from '../artifacts/contracts/src/transformers/WethTransformer.sol/WethTransformer.json';
+import * as IERC20Transformer from '../artifacts/contracts/src/transformers/IERC20Transformer.sol/IERC20Transformer.json';
+import * as LogMetadataTransformer from '../artifacts/contracts/src/transformers/LogMetadataTransformer.sol/LogMetadataTransformer.json';
+
+// Bridge Adapters
+import * as AvalancheBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/AvalancheBridgeAdapter.sol/AvalancheBridgeAdapter.json';
+import * as BSCBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/BSCBridgeAdapter.sol/BSCBridgeAdapter.json';
+import * as BaseBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/BaseBridgeAdapter.sol/BaseBridgeAdapter.json';
+import * as CeloBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/CeloBridgeAdapter.sol/CeloBridgeAdapter.json';
+import * as EthereumBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/EthereumBridgeAdapter.sol/EthereumBridgeAdapter.json';
+import * as FantomBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/FantomBridgeAdapter.sol/FantomBridgeAdapter.json';
+import * as OptimismBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/OptimismBridgeAdapter.sol/OptimismBridgeAdapter.json';
+import * as PolygonBridgeAdapter from '../artifacts/contracts/src/transformers/bridges/PolygonBridgeAdapter.sol/PolygonBridgeAdapter.json';
+
+// Fee Collectors
+import * as FeeCollector from '../artifacts/contracts/src/external/FeeCollector.sol/FeeCollector.json';
+import * as FeeCollectorController from '../artifacts/contracts/src/external/FeeCollectorController.sol/FeeCollectorController.json';
+
+/**
+ * Hardhat artifacts 对象
+ * 所有 artifacts 都使用 Hardhat JSON 格式
+ */
 export const artifacts = {
-    IZeroEx: IZeroEx as ContractArtifact,
-    ZeroEx: ZeroEx as ContractArtifact,
-    FullMigration: FullMigration as ContractArtifact,
-    InitialMigration: InitialMigration as ContractArtifact,
-    IFlashWallet: IFlashWallet as ContractArtifact,
-    IERC20Transformer: IERC20Transformer as ContractArtifact,
-    IOwnableFeature: IOwnableFeature as ContractArtifact,
-    ISimpleFunctionRegistryFeature: ISimpleFunctionRegistryFeature as ContractArtifact,
-    ITransformERC20Feature: ITransformERC20Feature as ContractArtifact,
-    FillQuoteTransformer: FillQuoteTransformer as ContractArtifact,
-    PayTakerTransformer: PayTakerTransformer as ContractArtifact,
-    PositiveSlippageFeeTransformer: PositiveSlippageFeeTransformer as ContractArtifact,
-    WethTransformer: WethTransformer as ContractArtifact,
-    OwnableFeature: OwnableFeature as ContractArtifact,
-    SimpleFunctionRegistryFeature: SimpleFunctionRegistryFeature as ContractArtifact,
-    TransformERC20Feature: TransformERC20Feature as ContractArtifact,
-    AffiliateFeeTransformer: AffiliateFeeTransformer as ContractArtifact,
-    MetaTransactionsFeature: MetaTransactionsFeature as ContractArtifact,
-    LogMetadataTransformer: LogMetadataTransformer as ContractArtifact,
-    LiquidityProviderFeature: LiquidityProviderFeature as ContractArtifact,
-    ILiquidityProviderFeature: ILiquidityProviderFeature as ContractArtifact,
-    NativeOrdersFeature: NativeOrdersFeature as ContractArtifact,
-    INativeOrdersFeature: INativeOrdersFeature as ContractArtifact,
-    FeeCollectorController: FeeCollectorController as ContractArtifact,
-    FeeCollector: FeeCollector as ContractArtifact,
-    CurveLiquidityProvider: CurveLiquidityProvider as ContractArtifact,
-    BatchFillNativeOrdersFeature: BatchFillNativeOrdersFeature as ContractArtifact,
-    IBatchFillNativeOrdersFeature: IBatchFillNativeOrdersFeature as ContractArtifact,
-    MultiplexFeature: MultiplexFeature as ContractArtifact,
-    IMultiplexFeature: IMultiplexFeature as ContractArtifact,
-    OtcOrdersFeature: OtcOrdersFeature as ContractArtifact,
-    IOtcOrdersFeature: IOtcOrdersFeature as ContractArtifact,
-    AvalancheBridgeAdapter: AvalancheBridgeAdapter as ContractArtifact,
-    BaseGoerliBridgeAdapter: BaseGoerliBridgeAdapter as ContractArtifact,
-    BaseBridgeAdapter: BaseBridgeAdapter as ContractArtifact,
-    BSCBridgeAdapter: BSCBridgeAdapter as ContractArtifact,
-    CeloBridgeAdapter: CeloBridgeAdapter as ContractArtifact,
-    EthereumBridgeAdapter: EthereumBridgeAdapter as ContractArtifact,
-    FantomBridgeAdapter: FantomBridgeAdapter as ContractArtifact,
-    OptimismBridgeAdapter: OptimismBridgeAdapter as ContractArtifact,
-    PolygonBridgeAdapter: PolygonBridgeAdapter as ContractArtifact,
-    MetaTransactionsFeatureV2: MetaTransactionsFeatureV2 as ContractArtifact,
+    ZeroEx,
+    ZeroExOptimized,
+    IZeroEx,
+    FullMigration,
+    InitialMigration,
+    FlashWallet,
+    IFlashWallet,
+    BatchFillNativeOrdersFeature,
+    IBatchFillNativeOrdersFeature,
+    LiquidityProviderFeature,
+    ILiquidityProviderFeature,
+    MultiplexFeature,
+    IMultiplexFeature,
+    NativeOrdersFeature,
+    INativeOrdersFeature,
+    OtcOrdersFeature,
+    IOtcOrdersFeature,
+    OwnableFeature,
+    IOwnableFeature,
+    SimpleFunctionRegistryFeature,
+    ISimpleFunctionRegistryFeature,
+    TransformERC20Feature,
+    ITransformERC20Feature,
+    MetaTransactionsFeature,
+    MetaTransactionsFeatureV2,
+    AffiliateFeeTransformer,
+    FillQuoteTransformer,
+    PayTakerTransformer,
+    PositiveSlippageFeeTransformer,
+    WethTransformer,
+    IERC20Transformer,
+    LogMetadataTransformer,
+    AvalancheBridgeAdapter,
+    BSCBridgeAdapter,
+    BaseBridgeAdapter,
+    CeloBridgeAdapter,
+    EthereumBridgeAdapter,
+    FantomBridgeAdapter,
+    OptimismBridgeAdapter,
+    PolygonBridgeAdapter,
+    FeeCollector,
+    FeeCollectorController,
 };
