@@ -3,45 +3,25 @@
  * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
  * -----------------------------------------------------------------------------
  */
-export * from '../generated-wrappers/affiliate_fee_transformer';
-export * from '../generated-wrappers/avalanche_bridge_adapter';
-export * from '../generated-wrappers/b_s_c_bridge_adapter';
-export * from '../generated-wrappers/base_bridge_adapter';
-export * from '../generated-wrappers/base_goerli_bridge_adapter';
-export * from '../generated-wrappers/batch_fill_native_orders_feature';
-export * from '../generated-wrappers/celo_bridge_adapter';
-export * from '../generated-wrappers/curve_liquidity_provider';
-export * from '../generated-wrappers/ethereum_bridge_adapter';
-export * from '../generated-wrappers/fantom_bridge_adapter';
-export * from '../generated-wrappers/fee_collector';
-export * from '../generated-wrappers/fee_collector_controller';
-export * from '../generated-wrappers/fill_quote_transformer';
-export * from '../generated-wrappers/full_migration';
-export * from '../generated-wrappers/i_batch_fill_native_orders_feature';
-export * from '../generated-wrappers/i_erc20_transformer';
-export * from '../generated-wrappers/i_flash_wallet';
-export * from '../generated-wrappers/i_liquidity_provider_feature';
-export * from '../generated-wrappers/i_multiplex_feature';
-export * from '../generated-wrappers/i_native_orders_feature';
-export * from '../generated-wrappers/i_otc_orders_feature';
-export * from '../generated-wrappers/i_ownable_feature';
-export * from '../generated-wrappers/i_simple_function_registry_feature';
-export * from '../generated-wrappers/i_transform_erc20_feature';
-export * from '../generated-wrappers/i_zero_ex';
-export * from '../generated-wrappers/initial_migration';
-export * from '../generated-wrappers/liquidity_provider_feature';
-export * from '../generated-wrappers/log_metadata_transformer';
-export * from '../generated-wrappers/meta_transactions_feature';
-export * from '../generated-wrappers/meta_transactions_feature_v2';
-export * from '../generated-wrappers/multiplex_feature';
-export * from '../generated-wrappers/native_orders_feature';
-export * from '../generated-wrappers/optimism_bridge_adapter';
-export * from '../generated-wrappers/otc_orders_feature';
-export * from '../generated-wrappers/ownable_feature';
-export * from '../generated-wrappers/pay_taker_transformer';
-export * from '../generated-wrappers/polygon_bridge_adapter';
-export * from '../generated-wrappers/positive_slippage_fee_transformer';
-export * from '../generated-wrappers/simple_function_registry_feature';
-export * from '../generated-wrappers/transform_erc20_feature';
-export * from '../generated-wrappers/weth_transformer';
-export * from '../generated-wrappers/zero_ex';
+
+// Export TypeChain generated contract types
+export { ZeroEx } from './typechain-types/contracts/src/ZeroEx';
+export { ZeroExOptimized } from './typechain-types/contracts/src/ZeroExOptimized';
+export { FullMigration } from './typechain-types/contracts/src/migrations/FullMigration';
+export { InitialMigration } from './typechain-types/contracts/src/migrations/InitialMigration';
+
+// Export factory classes (flatten common ones to avoid deep import issues)
+export { ZeroEx__factory } from './typechain-types/factories/contracts/src';
+export {
+    InitialMigration__factory,
+    FullMigration__factory,
+} from './typechain-types/factories/contracts/src/migrations';
+export * from './typechain-types/factories/contracts/src/features';
+export * from './typechain-types/factories/contracts/test';
+export * from './typechain-types/factories/contracts/test/tokens';
+export * from './typechain-types/factories/contracts/src/external';
+export * from './typechain-types/factories/contracts/src/transformers';
+export * from './typechain-types/factories/contracts/src/transformers/bridges';
+
+// Export common types
+export * from './typechain-types/common';
