@@ -1,25 +1,51 @@
-import { blockchainTests, describe, expect } from '@0x/contracts-test-utils';
+import { expect } from 'chai';
+import { ethers } from 'hardhat';
 
 import { artifacts } from './artifacts';
 import { getRandomLimitOrder, getRandomRfqOrder } from './utils/orders';
-import { TestLibNativeOrderContract } from './wrappers';
+import { TestLibNativeOrder__factory } from '../src/typechain-types/factories/contracts/test';
+import type { TestLibNativeOrder } from '../src/typechain-types/contracts/test/TestLibNativeOrder';
 
-blockchainTests('LibLimitOrder tests', env => {
-    let testContract: TestLibNativeOrderContract;
+describe('LibLimitOrder tests', () => {
+    const env = {
+        provider: ethers.provider,
+        txDefaults: { from: '' as string },
+        getAccountAddressesAsync: async (): Promise<string[]> => (await ethers.getSigners()).map(s => s.address),
+    } as any;
+    let testContract: TestLibNativeOrder;
 
     before(async () => {
-        testContract = await TestLibNativeOrderContract.deployFrom0xArtifactAsync(
-            artifacts.TestLibNativeOrder,
-            env.provider,
-            env.txDefaults,
-            artifacts,
-        );
+        // 使用测试环境中的 provider 和账户
+        const accounts = await env.getAccountAddressesAsync();
+        env.txDefaults.from = accounts[0];
+        const signer = await env.provider.getSigner(accounts[0]);
+
+        const factory = new TestLibNativeOrder__factory(signer);
+        testContract = await factory.deploy();
+        await testContract.waitForDeployment();
     });
 
     describe('getLimitOrderStructHash()', () => {
         it('returns the correct hash', async () => {
             const order = getRandomLimitOrder();
-            const structHash = await testContract.getLimitOrderStructHash(order).callAsync();
+
+            // 转换订单字段为 ethers v6 兼容格式
+            const orderStruct = {
+                makerToken: order.makerToken,
+                takerToken: order.takerToken,
+                makerAmount: order.makerAmount.toString(),
+                takerAmount: order.takerAmount.toString(),
+                takerTokenFeeAmount: order.takerTokenFeeAmount.toString(),
+                maker: order.maker,
+                taker: order.taker,
+                sender: order.sender,
+                feeRecipient: order.feeRecipient,
+                pool: order.pool,
+                expiry: order.expiry.toString(),
+                salt: order.salt.toString(),
+            };
+
+            const structHash = await testContract.getLimitOrderStructHash(orderStruct);
             expect(structHash).to.eq(order.getStructHash());
         });
     });
@@ -27,7 +53,22 @@ blockchainTests('LibLimitOrder tests', env => {
     describe('getRfqOrderStructHash()', () => {
         it('returns the correct hash', async () => {
             const order = getRandomRfqOrder();
-            const structHash = await testContract.getRfqOrderStructHash(order).callAsync();
+
+            // 转换订单字段为 ethers v6 兼容格式
+            const orderStruct = {
+                makerToken: order.makerToken,
+                takerToken: order.takerToken,
+                makerAmount: order.makerAmount.toString(),
+                takerAmount: order.takerAmount.toString(),
+                maker: order.maker,
+                taker: order.taker, // RFQ 订单的 taker 字段
+                txOrigin: order.txOrigin,
+                pool: order.pool,
+                expiry: order.expiry.toString(),
+                salt: order.salt.toString(),
+            };
+
+            const structHash = await testContract.getRfqOrderStructHash(orderStruct);
             expect(structHash).to.eq(order.getStructHash());
         });
     });
