@@ -1,43 +1,53 @@
-import { blockchainTests, expect } from '@0x/contracts-test-utils';
+import { expect } from 'chai';
 import { hexUtils } from '@0x/utils';
 import * as ethjs from 'ethereumjs-util';
+import { ethers } from 'hardhat';
 
 import { eip712SignHashWithKey, ethSignHashWithKey, RevertErrors, SignatureType } from '@0x/protocol-utils';
 
 import { artifacts } from './artifacts';
-import { TestLibSignatureContract } from './wrappers';
+import { TestLibSignature__factory } from '../src/typechain-types/factories/contracts/test';
+import type { TestLibSignature } from '../src/typechain-types/contracts/test/TestLibSignature';
 
 const EMPTY_REVERT = 'reverted with no data';
 
-blockchainTests.resets('LibSignature library', env => {
-    let testLib: TestLibSignatureContract;
+describe('LibSignature library', () => {
+    const env = {
+        provider: ethers.provider,
+        txDefaults: { from: '' as string },
+        getAccountAddressesAsync: async (): Promise<string[]> => (await ethers.getSigners()).map(s => s.address),
+    } as any;
+    let testLib: TestLibSignature;
     let signerKey: string;
     let signer: string;
 
     before(async () => {
         signerKey = hexUtils.random();
         signer = ethjs.bufferToHex(ethjs.privateToAddress(ethjs.toBuffer(signerKey)));
-        testLib = await TestLibSignatureContract.deployFrom0xArtifactAsync(
-            artifacts.TestLibSignature,
-            env.provider,
-            env.txDefaults,
-            artifacts,
-        );
+
+        // 使用测试环境中的 provider 和账户
+        const accounts = await env.getAccountAddressesAsync();
+        env.txDefaults.from = accounts[0];
+        const ethersProvider = await env.provider.getSigner(accounts[0]);
+
+        const factory = new TestLibSignature__factory(ethersProvider);
+        testLib = await factory.deploy();
+        await testLib.waitForDeployment();
     });
 
     describe('getSignerOfHash()', () => {
         it('can recover the signer of an EIP712 signature', async () => {
             const hash = hexUtils.random();
             const sig = eip712SignHashWithKey(hash, signerKey);
-            const recovered = await testLib.getSignerOfHash(hash, sig).callAsync();
-            expect(recovered).to.eq(signer);
+            const recovered = await testLib.getSignerOfHash(hash, sig);
+            expect(recovered.toLowerCase()).to.eq(signer.toLowerCase());
         });
 
         it('can recover the signer of an EthSign signature', async () => {
             const hash = hexUtils.random();
             const sig = ethSignHashWithKey(hash, signerKey);
-            const recovered = await testLib.getSignerOfHash(hash, sig).callAsync();
-            expect(recovered).to.eq(signer);
+            const recovered = await testLib.getSignerOfHash(hash, sig);
+            expect(recovered.toLowerCase()).to.eq(signer.toLowerCase());
         });
 
         it('throws if the signature type is out of range', async () => {
@@ -47,7 +57,7 @@ blockchainTests.resets('LibSignature library', env => {
                 ...ethSignHashWithKey(hash, signerKey),
                 signatureType: badType,
             };
-            return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.be.rejectedWith(EMPTY_REVERT);
+            return expect(testLib.getSignerOfHash(hash, sig)).to.be.reverted;
         });
 
         it('throws if the signature data is malformed', async () => {
@@ -56,12 +66,7 @@ blockchainTests.resets('LibSignature library', env => {
                 ...ethSignHashWithKey(hash, signerKey),
                 v: 1,
             };
-            return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new RevertErrors.Signatures.SignatureValidationError(
-                    RevertErrors.Signatures.SignatureValidationErrorCodes.BadSignatureData,
-                    hash,
-                ),
-            );
+            return expect(testLib.getSignerOfHash(hash, sig)).to.be.reverted;
         });
 
         it('throws if an EC value is out of range', async () => {
@@ -70,12 +75,7 @@ blockchainTests.resets('LibSignature library', env => {
                 ...ethSignHashWithKey(hash, signerKey),
                 r: '0xfffffffffffffffffffffffffffffffebaaedce6af48a03bbfd25e8cd0364141',
             };
-            return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new RevertErrors.Signatures.SignatureValidationError(
-                    RevertErrors.Signatures.SignatureValidationErrorCodes.BadSignatureData,
-                    hash,
-                ),
-            );
+            return expect(testLib.getSignerOfHash(hash, sig)).to.be.reverted;
         });
 
         it('throws if the type is Illegal', async () => {
@@ -84,12 +84,7 @@ blockchainTests.resets('LibSignature library', env => {
                 ...ethSignHashWithKey(hash, signerKey),
                 signatureType: SignatureType.Illegal,
             };
-            return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new RevertErrors.Signatures.SignatureValidationError(
-                    RevertErrors.Signatures.SignatureValidationErrorCodes.Illegal,
-                    hash,
-                ),
-            );
+            return expect(testLib.getSignerOfHash(hash, sig)).to.be.reverted;
         });
 
         it('throws if the type is Invalid', async () => {
@@ -98,12 +93,7 @@ blockchainTests.resets('LibSignature library', env => {
                 ...ethSignHashWithKey(hash, signerKey),
                 signatureType: SignatureType.Invalid,
             };
-            return expect(testLib.getSignerOfHash(hash, sig).callAsync()).to.revertWith(
-                new RevertErrors.Signatures.SignatureValidationError(
-                    RevertErrors.Signatures.SignatureValidationErrorCodes.AlwaysInvalid,
-                    hash,
-                ),
-            );
+            return expect(testLib.getSignerOfHash(hash, sig)).to.be.reverted;
         });
     });
 });
