@@ -2,4 +2,4 @@ import * as _ from 'lodash';
 
 import { artifacts } from '../artifacts';
 
-export const abis = _.mapValues(artifacts, v => v.compilerOutput.abi);
+export const abis = _.mapValues(artifacts, v => v.abi);
