@@ -1,207 +1,229 @@
-/*
- * -----------------------------------------------------------------------------
- * Warning: This file is auto-generated by contracts-gen. Don't edit manually.
- * -----------------------------------------------------------------------------
+/* eslint-disable @typescript-eslint/no-unused-vars */
+/**
+ * 合约类型和工厂导出 - 支持所有测试所需的合约
  */
-export * from '../test/generated-wrappers/abstract_bridge_adapter';
-export * from '../test/generated-wrappers/affiliate_fee_transformer';
-export * from '../test/generated-wrappers/arbitrum_bridge_adapter';
-export * from '../test/generated-wrappers/avalanche_bridge_adapter';
-export * from '../test/generated-wrappers/b_s_c_bridge_adapter';
-export * from '../test/generated-wrappers/base_bridge_adapter';
-export * from '../test/generated-wrappers/base_goerli_bridge_adapter';
-export * from '../test/generated-wrappers/batch_fill_native_orders_feature';
-export * from '../test/generated-wrappers/bootstrap_feature';
-export * from '../test/generated-wrappers/bridge_protocols';
-export * from '../test/generated-wrappers/celo_bridge_adapter';
-export * from '../test/generated-wrappers/curve_liquidity_provider';
-export * from '../test/generated-wrappers/erc1155_orders_feature';
-export * from '../test/generated-wrappers/erc165_feature';
-export * from '../test/generated-wrappers/erc721_orders_feature';
-export * from '../test/generated-wrappers/ethereum_bridge_adapter';
-export * from '../test/generated-wrappers/fantom_bridge_adapter';
-export * from '../test/generated-wrappers/fee_collector';
-export * from '../test/generated-wrappers/fee_collector_controller';
-export * from '../test/generated-wrappers/fill_quote_transformer';
-export * from '../test/generated-wrappers/fixin_common';
-export * from '../test/generated-wrappers/fixin_e_i_p712';
-export * from '../test/generated-wrappers/fixin_erc1155_spender';
-export * from '../test/generated-wrappers/fixin_erc721_spender';
-export * from '../test/generated-wrappers/fixin_protocol_fees';
-export * from '../test/generated-wrappers/fixin_reentrancy_guard';
-export * from '../test/generated-wrappers/fixin_token_spender';
-export * from '../test/generated-wrappers/flash_wallet';
-export * from '../test/generated-wrappers/full_migration';
-export * from '../test/generated-wrappers/fund_recovery_feature';
-export * from '../test/generated-wrappers/i_batch_fill_native_orders_feature';
-export * from '../test/generated-wrappers/i_bootstrap_feature';
-export * from '../test/generated-wrappers/i_bridge_adapter';
-export * from '../test/generated-wrappers/i_erc1155_orders_feature';
-export * from '../test/generated-wrappers/i_erc1155_token';
-export * from '../test/generated-wrappers/i_erc165_feature';
-export * from '../test/generated-wrappers/i_erc20_bridge';
-export * from '../test/generated-wrappers/i_erc20_transformer';
-export * from '../test/generated-wrappers/i_erc721_orders_feature';
-export * from '../test/generated-wrappers/i_erc721_token';
-export * from '../test/generated-wrappers/i_feature';
-export * from '../test/generated-wrappers/i_fee_recipient';
-export * from '../test/generated-wrappers/i_flash_wallet';
-export * from '../test/generated-wrappers/i_fund_recovery_feature';
-export * from '../test/generated-wrappers/i_liquidity_provider';
-export * from '../test/generated-wrappers/i_liquidity_provider_feature';
-export * from '../test/generated-wrappers/i_liquidity_provider_sandbox';
-export * from '../test/generated-wrappers/i_meta_transactions_feature';
-export * from '../test/generated-wrappers/i_meta_transactions_feature_v2';
-export * from '../test/generated-wrappers/i_mooniswap_pool';
-export * from '../test/generated-wrappers/i_multiplex_feature';
-export * from '../test/generated-wrappers/i_native_orders_events';
-export * from '../test/generated-wrappers/i_native_orders_feature';
-export * from '../test/generated-wrappers/i_otc_orders_feature';
-export * from '../test/generated-wrappers/i_ownable_feature';
-export * from '../test/generated-wrappers/i_pancake_swap_feature';
-export * from '../test/generated-wrappers/i_property_validator';
-export * from '../test/generated-wrappers/i_simple_function_registry_feature';
-export * from '../test/generated-wrappers/i_staking';
-export * from '../test/generated-wrappers/i_taker_callback';
-export * from '../test/generated-wrappers/i_test_simple_function_registry_feature';
-export * from '../test/generated-wrappers/i_token_spender_feature';
-export * from '../test/generated-wrappers/i_transform_erc20_feature';
-export * from '../test/generated-wrappers/i_uniswap_feature';
-export * from '../test/generated-wrappers/i_uniswap_v2_pair';
-export * from '../test/generated-wrappers/i_uniswap_v3_feature';
-export * from '../test/generated-wrappers/i_uniswap_v3_pool';
-export * from '../test/generated-wrappers/i_zero_ex';
-export * from '../test/generated-wrappers/initial_migration';
-export * from '../test/generated-wrappers/lib_bootstrap';
-export * from '../test/generated-wrappers/lib_common_rich_errors';
-export * from '../test/generated-wrappers/lib_erc1155_orders_storage';
-export * from '../test/generated-wrappers/lib_erc20_transformer';
-export * from '../test/generated-wrappers/lib_erc721_orders_storage';
-export * from '../test/generated-wrappers/lib_fee_collector';
-export * from '../test/generated-wrappers/lib_liquidity_provider_rich_errors';
-export * from '../test/generated-wrappers/lib_meta_transactions_rich_errors';
-export * from '../test/generated-wrappers/lib_meta_transactions_storage';
-export * from '../test/generated-wrappers/lib_migrate';
-export * from '../test/generated-wrappers/lib_n_f_t_order';
-export * from '../test/generated-wrappers/lib_n_f_t_orders_rich_errors';
-export * from '../test/generated-wrappers/lib_native_order';
-export * from '../test/generated-wrappers/lib_native_orders_rich_errors';
-export * from '../test/generated-wrappers/lib_native_orders_storage';
-export * from '../test/generated-wrappers/lib_otc_orders_storage';
-export * from '../test/generated-wrappers/lib_ownable_rich_errors';
-export * from '../test/generated-wrappers/lib_ownable_storage';
-export * from '../test/generated-wrappers/lib_proxy_rich_errors';
-export * from '../test/generated-wrappers/lib_proxy_storage';
-export * from '../test/generated-wrappers/lib_reentrancy_guard_storage';
-export * from '../test/generated-wrappers/lib_signature';
-export * from '../test/generated-wrappers/lib_signature_rich_errors';
-export * from '../test/generated-wrappers/lib_simple_function_registry_rich_errors';
-export * from '../test/generated-wrappers/lib_simple_function_registry_storage';
-export * from '../test/generated-wrappers/lib_storage';
-export * from '../test/generated-wrappers/lib_transform_erc20_rich_errors';
-export * from '../test/generated-wrappers/lib_transform_erc20_storage';
-export * from '../test/generated-wrappers/lib_wallet_rich_errors';
-export * from '../test/generated-wrappers/liquidity_provider_feature';
-export * from '../test/generated-wrappers/liquidity_provider_sandbox';
-export * from '../test/generated-wrappers/log_metadata_transformer';
-export * from '../test/generated-wrappers/meta_transactions_feature';
-export * from '../test/generated-wrappers/meta_transactions_feature_v2';
-export * from '../test/generated-wrappers/mixin_aave_v2';
-export * from '../test/generated-wrappers/mixin_balancer';
-export * from '../test/generated-wrappers/mixin_balancer_v2_batch';
-export * from '../test/generated-wrappers/mixin_bancor_v3';
-export * from '../test/generated-wrappers/mixin_compound';
-export * from '../test/generated-wrappers/mixin_crypto_com';
-export * from '../test/generated-wrappers/mixin_curve';
-export * from '../test/generated-wrappers/mixin_curve_v2';
-export * from '../test/generated-wrappers/mixin_dodo';
-export * from '../test/generated-wrappers/mixin_dodo_v2';
-export * from '../test/generated-wrappers/mixin_g_m_x';
-export * from '../test/generated-wrappers/mixin_kyber_dmm';
-export * from '../test/generated-wrappers/mixin_lido';
-export * from '../test/generated-wrappers/mixin_maker_p_s_m';
-export * from '../test/generated-wrappers/mixin_mooniswap';
-export * from '../test/generated-wrappers/mixin_nerve';
-export * from '../test/generated-wrappers/mixin_platypus';
-export * from '../test/generated-wrappers/mixin_solidly';
-export * from '../test/generated-wrappers/mixin_synthetix';
-export * from '../test/generated-wrappers/mixin_uniswap';
-export * from '../test/generated-wrappers/mixin_uniswap_v2';
-export * from '../test/generated-wrappers/mixin_uniswap_v3';
-export * from '../test/generated-wrappers/mixin_zero_ex_bridge';
-export * from '../test/generated-wrappers/mooniswap_liquidity_provider';
-export * from '../test/generated-wrappers/multiplex_feature';
-export * from '../test/generated-wrappers/multiplex_liquidity_provider';
-export * from '../test/generated-wrappers/multiplex_otc';
-export * from '../test/generated-wrappers/multiplex_rfq';
-export * from '../test/generated-wrappers/multiplex_transform_erc20';
-export * from '../test/generated-wrappers/multiplex_uniswap_v2';
-export * from '../test/generated-wrappers/multiplex_uniswap_v3';
-export * from '../test/generated-wrappers/n_f_t_orders';
-export * from '../test/generated-wrappers/native_orders_cancellation';
-export * from '../test/generated-wrappers/native_orders_feature';
-export * from '../test/generated-wrappers/native_orders_info';
-export * from '../test/generated-wrappers/native_orders_protocol_fees';
-export * from '../test/generated-wrappers/native_orders_settlement';
-export * from '../test/generated-wrappers/optimism_bridge_adapter';
-export * from '../test/generated-wrappers/otc_orders_feature';
-export * from '../test/generated-wrappers/ownable_feature';
-export * from '../test/generated-wrappers/pancake_swap_feature';
-export * from '../test/generated-wrappers/pay_taker_transformer';
-export * from '../test/generated-wrappers/permissionless_transformer_deployer';
-export * from '../test/generated-wrappers/polygon_bridge_adapter';
-export * from '../test/generated-wrappers/positive_slippage_fee_transformer';
-export * from '../test/generated-wrappers/simple_function_registry_feature';
-export * from '../test/generated-wrappers/test_curve';
-export * from '../test/generated-wrappers/test_delegate_caller';
-export * from '../test/generated-wrappers/test_fee_collector_controller';
-export * from '../test/generated-wrappers/test_fee_recipient';
-export * from '../test/generated-wrappers/test_fill_quote_transformer_bridge';
-export * from '../test/generated-wrappers/test_fill_quote_transformer_exchange';
-export * from '../test/generated-wrappers/test_fill_quote_transformer_host';
-export * from '../test/generated-wrappers/test_fixin_protocol_fees';
-export * from '../test/generated-wrappers/test_fixin_token_spender';
-export * from '../test/generated-wrappers/test_full_migration';
-export * from '../test/generated-wrappers/test_initial_migration';
-export * from '../test/generated-wrappers/test_lib_native_order';
-export * from '../test/generated-wrappers/test_lib_signature';
-export * from '../test/generated-wrappers/test_liquidity_provider';
-export * from '../test/generated-wrappers/test_meta_transactions_native_orders_feature';
-export * from '../test/generated-wrappers/test_meta_transactions_transform_erc20_feature';
-export * from '../test/generated-wrappers/test_migrator';
-export * from '../test/generated-wrappers/test_mint_token_erc20_transformer';
-export * from '../test/generated-wrappers/test_mintable_erc1155_token';
-export * from '../test/generated-wrappers/test_mintable_erc20_token';
-export * from '../test/generated-wrappers/test_mintable_erc721_token';
-export * from '../test/generated-wrappers/test_mooniswap';
-export * from '../test/generated-wrappers/test_n_f_t_order_presigner';
-export * from '../test/generated-wrappers/test_native_orders_feature';
-export * from '../test/generated-wrappers/test_no_eth_recipient';
-export * from '../test/generated-wrappers/test_order_signer_registry_with_contract_wallet';
-export * from '../test/generated-wrappers/test_permissionless_transformer_deployer_suicidal';
-export * from '../test/generated-wrappers/test_permissionless_transformer_deployer_transformer';
-export * from '../test/generated-wrappers/test_property_validator';
-export * from '../test/generated-wrappers/test_rfq_origin_registration';
-export * from '../test/generated-wrappers/test_simple_function_registry_feature_impl1';
-export * from '../test/generated-wrappers/test_simple_function_registry_feature_impl2';
-export * from '../test/generated-wrappers/test_staking';
-export * from '../test/generated-wrappers/test_token_spender_erc20_token';
-export * from '../test/generated-wrappers/test_transform_erc20';
-export * from '../test/generated-wrappers/test_transformer_base';
-export * from '../test/generated-wrappers/test_transformer_deployer_transformer';
-export * from '../test/generated-wrappers/test_transformer_host';
-export * from '../test/generated-wrappers/test_uniswap_v2_factory';
-export * from '../test/generated-wrappers/test_uniswap_v2_pool';
-export * from '../test/generated-wrappers/test_uniswap_v3_factory';
-export * from '../test/generated-wrappers/test_uniswap_v3_feature';
-export * from '../test/generated-wrappers/test_uniswap_v3_pool';
-export * from '../test/generated-wrappers/test_weth';
-export * from '../test/generated-wrappers/test_weth_transformer_host';
-export * from '../test/generated-wrappers/transform_erc20_feature';
-export * from '../test/generated-wrappers/transformer';
-export * from '../test/generated-wrappers/transformer_deployer';
-export * from '../test/generated-wrappers/uniswap_feature';
-export * from '../test/generated-wrappers/uniswap_v3_feature';
-export * from '../test/generated-wrappers/weth_transformer';
-export * from '../test/generated-wrappers/zero_ex';
-export * from '../test/generated-wrappers/zero_ex_optimized';
+
+// 核心合约类型导出
+export type { ZeroEx as ZeroExContract } from '../src/typechain-types/contracts/src/ZeroEx';
+export type { IZeroEx as IZeroExContract } from '../src/typechain-types/contracts/src/IZeroEx';
+
+// 测试合约类型导出 - 按字母顺序排列
+export type { TestFullMigration as TestFullMigrationContract } from '../src/typechain-types/contracts/test/TestFullMigration';
+export type { TestInitialMigration as TestInitialMigrationContract } from '../src/typechain-types/contracts/test/TestInitialMigration';
+export type { TestFixinProtocolFees as TestFixinProtocolFeesContract } from '../src/typechain-types/contracts/test/TestFixinProtocolFees';
+export type { TestFixinTokenSpender as TestFixinTokenSpenderContract } from '../src/typechain-types/contracts/test/TestFixinTokenSpender';
+export type { TestMigrator as TestMigratorContract } from '../src/typechain-types/contracts/test/TestMigrator';
+export type { TestStaking as TestStakingContract } from '../src/typechain-types/contracts/test/TestStaking';
+
+// 代币相关测试合约
+export type { TestMintableERC20Token as TestMintableERC20TokenContract } from '../src/typechain-types/contracts/test/tokens/TestMintableERC20Token';
+export type { TestMintableERC721Token as TestMintableERC721TokenContract } from '../src/typechain-types/contracts/test/tokens/TestMintableERC721Token';
+export type { TestMintableERC1155Token as TestMintableERC1155TokenContract } from '../src/typechain-types/contracts/test/tokens/TestMintableERC1155Token';
+export type { TestTokenSpenderERC20Token as TestTokenSpenderERC20TokenContract } from '../src/typechain-types/contracts/test/tokens/TestTokenSpenderERC20Token';
+export type { TestWeth as TestWethContract } from '../src/typechain-types/contracts/test/tokens/TestWeth';
+export type { DummyERC20Token as DummyERC20TokenContract } from '../src/typechain-types/contracts/src/external/DummyERC20Token';
+
+// Feature 接口类型
+export type { IFundRecoveryFeature as IFundRecoveryFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IFundRecoveryFeature';
+export type { IBatchFillNativeOrdersFeature as IBatchFillNativeOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IBatchFillNativeOrdersFeature';
+export type { IOtcOrdersFeature as IOtcOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IOtcOrdersFeature';
+export type { IERC721OrdersFeature as IERC721OrdersFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IERC721OrdersFeature';
+export type { IERC1155OrdersFeature as IERC1155OrdersFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IERC1155OrdersFeature';
+export type { INativeOrdersFeature as INativeOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/INativeOrdersFeature';
+export type { ISimpleFunctionRegistryFeature as ISimpleFunctionRegistryFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/ISimpleFunctionRegistryFeature';
+export type { IOwnableFeature as IOwnableFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IOwnableFeature';
+export type { ITransformERC20Feature as ITransformERC20FeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/ITransformERC20Feature';
+export type { IMetaTransactionsFeature as IMetaTransactionsFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IMetaTransactionsFeature';
+export type { ILiquidityProviderFeature as ILiquidityProviderFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/ILiquidityProviderFeature';
+export type { IUniswapV3Feature as IUniswapV3FeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IUniswapV3Feature';
+export type { IMultiplexFeature as IMultiplexFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IMultiplexFeature';
+export type { IBootstrapFeature as IBootstrapFeatureContract } from '../src/typechain-types/contracts/src/features/interfaces/IBootstrapFeature';
+
+// Feature 实现类型
+export type { FundRecoveryFeature as FundRecoveryFeatureContract } from '../src/typechain-types/contracts/src/features/FundRecoveryFeature';
+export type { BatchFillNativeOrdersFeature as BatchFillNativeOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/BatchFillNativeOrdersFeature';
+export type { OtcOrdersFeature as OtcOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/OtcOrdersFeature';
+export type { ERC721OrdersFeature as ERC721OrdersFeatureContract } from '../src/typechain-types/contracts/src/features/ERC721OrdersFeature';
+export type { ERC1155OrdersFeature as ERC1155OrdersFeatureContract } from '../src/typechain-types/contracts/src/features/ERC1155OrdersFeature';
+export type { NativeOrdersFeature as NativeOrdersFeatureContract } from '../src/typechain-types/contracts/src/features/NativeOrdersFeature';
+export type { SimpleFunctionRegistryFeature as SimpleFunctionRegistryFeatureContract } from '../src/typechain-types/contracts/src/features/SimpleFunctionRegistryFeature';
+export type { OwnableFeature as OwnableFeatureContract } from '../src/typechain-types/contracts/src/features/OwnableFeature';
+export type { TransformERC20Feature as TransformERC20FeatureContract } from '../src/typechain-types/contracts/src/features/TransformERC20Feature';
+export type { MetaTransactionsFeature as MetaTransactionsFeatureContract } from '../src/typechain-types/contracts/src/features/MetaTransactionsFeature';
+export type { LiquidityProviderFeature as LiquidityProviderFeatureContract } from '../src/typechain-types/contracts/src/features/LiquidityProviderFeature';
+export type { UniswapV3Feature as UniswapV3FeatureContract } from '../src/typechain-types/contracts/src/features/UniswapV3Feature';
+export type { MultiplexFeature as MultiplexFeatureContract } from '../src/typechain-types/contracts/src/features/MultiplexFeature';
+
+// Feature 测试合约类型
+export type { TestSimpleFunctionRegistryFeatureImpl1 as TestSimpleFunctionRegistryFeatureImpl1Contract } from '../src/typechain-types/contracts/test/TestSimpleFunctionRegistryFeatureImpl1';
+export type { TestSimpleFunctionRegistryFeatureImpl2 as TestSimpleFunctionRegistryFeatureImpl2Contract } from '../src/typechain-types/contracts/test/TestSimpleFunctionRegistryFeatureImpl2';
+export type { TestMetaTransactionsTransformERC20Feature as TestMetaTransactionsTransformERC20FeatureContract } from '../src/typechain-types/contracts/test/TestMetaTransactionsTransformERC20Feature';
+export type { TestMetaTransactionsNativeOrdersFeature as TestMetaTransactionsNativeOrdersFeatureContract } from '../src/typechain-types/contracts/test/TestMetaTransactionsNativeOrdersFeature';
+export type { TestTransformERC20 as TestTransformERC20Contract } from '../src/typechain-types/contracts/test/TestTransformERC20';
+export type { TestNativeOrdersFeature as TestNativeOrdersFeatureContract } from '../src/typechain-types/contracts/test/TestNativeOrdersFeature';
+
+// Transformer 相关类型
+export type { WethTransformer as WethTransformerContract } from '../src/typechain-types/contracts/src/transformers/WethTransformer';
+export type { PayTakerTransformer as PayTakerTransformerContract } from '../src/typechain-types/contracts/src/transformers/PayTakerTransformer';
+export type { FillQuoteTransformer as FillQuoteTransformerContract } from '../src/typechain-types/contracts/src/transformers/FillQuoteTransformer';
+export type { TestTransformerHost as TestTransformerHostContract } from '../src/typechain-types/contracts/test/TestTransformerHost';
+export type { TestWethTransformerHost as TestWethTransformerHostContract } from '../src/typechain-types/contracts/test/TestWethTransformerHost';
+export type { TestFillQuoteTransformerHost as TestFillQuoteTransformerHostContract } from '../src/typechain-types/contracts/test/TestFillQuoteTransformerHost';
+export type { TestTransformerBase as TestTransformerBaseContract } from '../src/typechain-types/contracts/test/TestTransformerBase';
+export type { TestDelegateCaller as TestDelegateCallerContract } from '../src/typechain-types/contracts/test/TestDelegateCaller';
+export type { TestMintTokenERC20Transformer as TestMintTokenERC20TransformerContract } from '../src/typechain-types/contracts/test/TestMintTokenERC20Transformer';
+export type { TestFillQuoteTransformerExchange as TestFillQuoteTransformerExchangeContract } from '../src/typechain-types/contracts/test/TestFillQuoteTransformerExchange';
+export type { TestFillQuoteTransformerBridge as TestFillQuoteTransformerBridgeContract } from '../src/typechain-types/contracts/test/TestFillQuoteTransformerBridge';
+export type { EthereumBridgeAdapter as EthereumBridgeAdapterContract } from '../src/typechain-types/contracts/src/transformers/bridges/EthereumBridgeAdapter';
+
+// 部署器相关类型
+export type { TransformerDeployer as TransformerDeployerContract } from '../src/typechain-types/contracts/src/external/TransformerDeployer';
+export type { TestTransformerDeployerTransformer as TestTransformerDeployerTransformerContract } from '../src/typechain-types/contracts/test/TestTransformerDeployerTransformer';
+export type { PermissionlessTransformerDeployer as PermissionlessTransformerDeployerContract } from '../src/typechain-types/contracts/src/external/PermissionlessTransformerDeployer';
+export type { TestPermissionlessTransformerDeployerTransformer as TestPermissionlessTransformerDeployerTransformerContract } from '../src/typechain-types/contracts/test/TestPermissionlessTransformerDeployerTransformer';
+
+// 流动性提供商相关类型
+export type { LiquidityProviderSandbox as LiquidityProviderSandboxContract } from '../src/typechain-types/contracts/src/external/LiquidityProviderSandbox';
+export type { TestLiquidityProvider as TestLiquidityProviderContract } from '../src/typechain-types/contracts/test/TestLiquidityProvider';
+export type { CurveLiquidityProvider as CurveLiquidityProviderContract } from '../src/typechain-types/contracts/src/liquidity-providers/CurveLiquidityProvider';
+export type { MooniswapLiquidityProvider as MooniswapLiquidityProviderContract } from '../src/typechain-types/contracts/src/liquidity-providers/MooniswapLiquidityProvider';
+export type { TestCurve as TestCurveContract } from '../src/typechain-types/contracts/test/integration/TestCurve';
+export type { TestMooniswap as TestMooniswapContract } from '../src/typechain-types/contracts/test/integration/TestMooniswap';
+
+// 集成测试合约类型
+export type { TestUniswapV2Factory as TestUniswapV2FactoryContract } from '../src/typechain-types/contracts/test/integration/TestUniswapV2Factory';
+export type { TestUniswapV3Factory as TestUniswapV3FactoryContract } from '../src/typechain-types/contracts/test/integration/TestUniswapV3Factory';
+export type { TestNoEthRecipient as TestNoEthRecipientContract } from '../src/typechain-types/contracts/test/TestNoEthRecipient';
+
+// 费用和控制器类型
+export type { FeeCollectorController as FeeCollectorControllerContract } from '../src/typechain-types/contracts/src/external/FeeCollectorController';
+
+// 注册和验证器类型
+export type { TestRfqOriginRegistration as TestRfqOriginRegistrationContract } from '../src/typechain-types/contracts/test/TestRfqOriginRegistration';
+export type { TestOrderSignerRegistryWithContractWallet as TestOrderSignerRegistryWithContractWalletContract } from '../src/typechain-types/contracts/test/TestOrderSignerRegistryWithContractWallet';
+export type { TestFeeRecipient as TestFeeRecipientContract } from '../src/typechain-types/contracts/test/TestFeeRecipient';
+export type { TestPropertyValidator as TestPropertyValidatorContract } from '../src/typechain-types/contracts/test/TestPropertyValidator';
+export type { TestNFTOrderPresigner as TestNFTOrderPresignerContract } from '../src/typechain-types/contracts/test/TestNFTOrderPresigner';
+
+// 迁移相关类型
+export type { InitialMigration as InitialMigrationContract } from '../src/typechain-types/contracts/src/migrations/InitialMigration';
+
+// 库合约类型
+export type { TestLibNativeOrder as TestLibNativeOrderContract } from '../src/typechain-types/contracts/test/TestLibNativeOrder';
+export type { TestLibSignature as TestLibSignatureContract } from '../src/typechain-types/contracts/test/TestLibSignature';
+
+// ================================
+// 工厂类型导出
+// ================================
+
+// 核心合约工厂
+export { ZeroEx__factory, IZeroEx__factory } from '../src/typechain-types/factories/contracts/src';
+
+// 测试合约工厂
+export {
+    TestFullMigration__factory,
+    TestInitialMigration__factory,
+    TestFixinProtocolFees__factory,
+    TestFixinTokenSpender__factory,
+    TestMigrator__factory,
+    TestStaking__factory,
+    TestLibNativeOrder__factory,
+    TestLibSignature__factory,
+    TestTransformerHost__factory,
+    TestWethTransformerHost__factory,
+    TestFillQuoteTransformerHost__factory,
+    TestTransformerBase__factory,
+    TestDelegateCaller__factory,
+    TestMintTokenERC20Transformer__factory,
+    TestFillQuoteTransformerExchange__factory,
+    TestFillQuoteTransformerBridge__factory,
+    TestTransformerDeployerTransformer__factory,
+    TestPermissionlessTransformerDeployerTransformer__factory,
+    TestLiquidityProvider__factory,
+    TestRfqOriginRegistration__factory,
+    TestOrderSignerRegistryWithContractWallet__factory,
+    TestFeeRecipient__factory,
+    TestPropertyValidator__factory,
+    TestNFTOrderPresigner__factory,
+    TestSimpleFunctionRegistryFeatureImpl1__factory,
+    TestSimpleFunctionRegistryFeatureImpl2__factory,
+    TestMetaTransactionsTransformERC20Feature__factory,
+    TestMetaTransactionsNativeOrdersFeature__factory,
+    TestTransformERC20__factory,
+    TestNativeOrdersFeature__factory,
+} from '../src/typechain-types/factories/contracts/test';
+
+// 代币测试合约工厂
+export {
+    TestMintableERC20Token__factory,
+    TestTokenSpenderERC20Token__factory,
+    TestWeth__factory,
+} from '../src/typechain-types/factories/contracts/test/tokens';
+
+export { TestMintableERC721Token__factory } from '../src/typechain-types/factories/contracts/test/tokens/TestMintableERC721Token.sol';
+export { TestMintableERC1155Token__factory } from '../src/typechain-types/factories/contracts/test/tokens/TestMintableERC1155Token.sol';
+
+// Feature 工厂
+export {
+    FundRecoveryFeature__factory,
+    BatchFillNativeOrdersFeature__factory,
+    OtcOrdersFeature__factory,
+    ERC721OrdersFeature__factory,
+    ERC1155OrdersFeature__factory,
+    NativeOrdersFeature__factory,
+    SimpleFunctionRegistryFeature__factory,
+    OwnableFeature__factory,
+    TransformERC20Feature__factory,
+    MetaTransactionsFeature__factory,
+    LiquidityProviderFeature__factory,
+    UniswapV3Feature__factory,
+    MultiplexFeature__factory,
+} from '../src/typechain-types/factories/contracts/src/features';
+
+// Transformer 工厂
+export {
+    WethTransformer__factory,
+    PayTakerTransformer__factory,
+    FillQuoteTransformer__factory,
+} from '../src/typechain-types/factories/contracts/src/transformers';
+
+// 桥接适配器工厂
+export { EthereumBridgeAdapter__factory } from '../src/typechain-types/factories/contracts/src/transformers/bridges';
+
+// 部署器工厂
+export { TransformerDeployer__factory } from '../src/typechain-types/factories/contracts/src/external/TransformerDeployer.sol';
+export {
+    PermissionlessTransformerDeployer__factory,
+    LiquidityProviderSandbox__factory,
+    FeeCollectorController__factory,
+} from '../src/typechain-types/factories/contracts/src/external';
+
+// 流动性提供商工厂
+export {
+    CurveLiquidityProvider__factory,
+    MooniswapLiquidityProvider__factory,
+} from '../src/typechain-types/factories/contracts/src/liquidity-providers';
+
+// 集成测试工厂
+export {
+    TestCurve__factory,
+    TestMooniswap__factory,
+    TestUniswapV2Factory__factory,
+    TestUniswapV3Factory__factory,
+} from '../src/typechain-types/factories/contracts/test/integration';
+
+// 外部合约工厂
+export { DummyERC20Token__factory } from '../src/typechain-types/factories/contracts/src/external';
+
+// 迁移工厂
+export { InitialMigration__factory } from '../src/typechain-types/factories/contracts/src/migrations';
+
+// 事件类型 - 用于测试中的事件验证
+export interface TransformerDeployerEvents {
+    Deployed: any;
+    Killed: any;
+}
+
+export interface PermissionlessTransformerDeployerEvents {
+    Deployed: any;
+}
