@@ -2,9 +2,9 @@
 $(document).ready(function () {
     $('a[href^="http://"], a[href^="https://"]').not('a[class*=internal]').attr('target', '_blank');
 
-    $(".researchPdf").height($( document ).height() * 0.8);
- });
+    $('.researchPdf').height($(document).height() * 0.8);
+});
 
-$( window ).resize(function() {
-    $(".researchPdf").height($( document ).height() * 0.8);
+$(window).resize(function () {
+    $('.researchPdf').height($(document).height() * 0.8);
 });
