@@ -1,61 +1,26 @@
-import * as AssetProxyOwner from '../artifacts/AssetProxyOwner.json';
-import * as Broker from '../artifacts/Broker.json';
-import * as Coordinator from '../artifacts/Coordinator.json';
-import * as CoordinatorRegistry from '../artifacts/CoordinatorRegistry.json';
-import * as DevUtils from '../artifacts/DevUtils.json';
 import * as DummyERC20Token from '../artifacts/DummyERC20Token.json';
-import * as DummyERC721Token from '../artifacts/DummyERC721Token.json';
-import * as ERC1155Mintable from '../artifacts/ERC1155Mintable.json';
-import * as ERC1155Proxy from '../artifacts/ERC1155Proxy.json';
-import * as ERC20BridgeProxy from '../artifacts/ERC20BridgeProxy.json';
-import * as ERC20Proxy from '../artifacts/ERC20Proxy.json';
 import * as ERC20Token from '../artifacts/ERC20Token.json';
-import * as ERC721Proxy from '../artifacts/ERC721Proxy.json';
-import * as ERC721Token from '../artifacts/ERC721Token.json';
-import * as Exchange from '../artifacts/Exchange.json';
-import * as Forwarder from '../artifacts/Forwarder.json';
-import * as GodsUnchainedValidator from '../artifacts/GodsUnchainedValidator.json';
-import * as IAssetProxy from '../artifacts/IAssetProxy.json';
-import * as ITransformERC20 from '../artifacts/ITransformERC20.json';
-import * as IValidator from '../artifacts/IValidator.json';
-import * as IWallet from '../artifacts/IWallet.json';
+import * as IERC20Token from '../artifacts/IERC20Token.json';
+import * as IEtherToken from '../artifacts/IEtherToken.json';
 import * as IZeroEx from '../artifacts/IZeroEx.json';
-import * as MultiAssetProxy from '../artifacts/MultiAssetProxy.json';
-import * as Staking from '../artifacts/Staking.json';
-import * as StakingProxy from '../artifacts/StakingProxy.json';
-import * as StaticCallProxy from '../artifacts/StaticCallProxy.json';
 import * as WETH9 from '../artifacts/WETH9.json';
-import * as ZRXToken from '../artifacts/ZRXToken.json';
-import * as ZrxVault from '../artifacts/ZrxVault.json';
+import * as ZRXWrappedToken from '../artifacts/ZRXWrappedToken.json';
+import * as LibERC20Token from '../artifacts/LibERC20Token.json';
+import * as UnlimitedAllowanceERC20Token from '../artifacts/UnlimitedAllowanceERC20Token.json';
 
+// Core contracts
 export {
-    AssetProxyOwner,
-    Broker,
-    DevUtils,
     DummyERC20Token,
-    DummyERC721Token,
-    ERC1155Mintable,
-    ERC1155Proxy,
-    ERC20BridgeProxy,
-    ERC20Proxy,
     ERC20Token,
-    ERC721Proxy,
-    ERC721Token,
-    Exchange,
-    Forwarder,
-    GodsUnchainedValidator,
-    IAssetProxy,
-    IValidator,
-    IWallet,
-    MultiAssetProxy,
-    StaticCallProxy,
-    WETH9,
-    ZRXToken,
-    Coordinator,
-    CoordinatorRegistry,
-    Staking,
-    StakingProxy,
-    ZrxVault,
-    ITransformERC20,
+    IERC20Token,
+    IEtherToken,
     IZeroEx,
+    WETH9,
+    ZRXWrappedToken,
+    LibERC20Token,
+    UnlimitedAllowanceERC20Token,
+    // 别名保持向后兼容性
+    ERC20Token as Token,
+    ZRXWrappedToken as ZRXToken,
+    UnlimitedAllowanceERC20Token as UnlimitedAllowanceToken,
 };
