@@ -1,14 +1,9 @@
-import { assert } from '@0x/assert';
 import { ContractAddresses } from '@0x/contract-addresses';
-import { AbiDecoder } from '@0x/utils';
-import { Web3Wrapper } from '@0x/web3-wrapper';
 import { SupportedProvider } from 'ethereum-types';
+import { ethers } from 'ethers';
+import { expect } from 'chai';
 
-import { ContractWrappersConfigSchema } from './contract_wrappers_config_schema';
-import { CoordinatorContract } from './generated-wrappers/coordinator';
-import { IZeroExContract } from './generated-wrappers/i_zero_ex';
-import { StakingContract } from './generated-wrappers/staking';
-import { WETH9Contract } from './generated-wrappers/weth9';
+import { WETH9__factory, IZeroEx__factory } from './typechain-types/factories';
 import { ContractWrappersConfig } from './types';
 import { _getDefaultContractAddresses } from './utils/contract_addresses';
 
@@ -21,20 +16,17 @@ export class ContractWrappers {
      */
     public contractAddresses: ContractAddresses;
     /**
-     * An instance of the WETH9Contract class containing methods for interacting with the
-     * WETH9 smart contract.
+     * WETH9 contract factory for creating contract instances
      */
-    public weth9: WETH9Contract;
+    public weth9Factory: typeof WETH9__factory;
     /**
-     * An instance of the StakingContract class containing methods for interacting with the Staking contracts.
+     * IZeroEx contract factory for creating contract instances
      */
-    public staking: StakingContract;
-    /**
-     * An instance of the IZeroExContract class containing methods for interacting with the Exchange Proxy.
-     */
-    public exchangeProxy: IZeroExContract;
+    public exchangeProxyFactory: typeof IZeroEx__factory;
 
-    private readonly _web3Wrapper: Web3Wrapper;
+    private readonly _ethersProvider: ethers.Provider;
+    private readonly _supportedProvider: SupportedProvider;
+
     /**
      * Instantiates a new ContractWrappers instance.
      * @param   supportedProvider    The Provider instance you would like the contract-wrappers library to use for interacting with
@@ -43,42 +35,55 @@ export class ContractWrappers {
      * @return  An instance of the ContractWrappers class.
      */
     constructor(supportedProvider: SupportedProvider, config: ContractWrappersConfig) {
-        assert.doesConformToSchema('config', config, ContractWrappersConfigSchema);
-        const txDefaults = {
-            gasPrice: config.gasPrice,
-        };
-        this._web3Wrapper = new Web3Wrapper(supportedProvider, txDefaults);
-        const contractsArray = [CoordinatorContract, StakingContract, WETH9Contract, IZeroExContract];
-        contractsArray.forEach(contract => {
-            this._web3Wrapper.abiDecoder.addABI(contract.ABI(), contract.contractName);
-        });
-        const contractAddresses =
-            config.contractAddresses === undefined
-                ? _getDefaultContractAddresses(config.chainId)
-                : config.contractAddresses;
-        this.weth9 = new WETH9Contract(contractAddresses.etherToken, this.getProvider());
-        this.staking = new StakingContract(contractAddresses.stakingProxy, this.getProvider());
-        this.exchangeProxy = new IZeroExContract(contractAddresses.exchangeProxy, this.getProvider());
-        this.contractAddresses = contractAddresses;
+        // Validate config using simple checks instead of JSON schema
+        expect(config).to.be.an('object');
+        expect(config.chainId).to.be.a('number');
+
+        this._supportedProvider = supportedProvider;
+        this.contractAddresses = config.contractAddresses || _getDefaultContractAddresses(config.chainId);
+
+        // Create ethers provider from the supported provider
+        // Note: We cast to unknown first to handle the type conversion safely
+        this._ethersProvider = new ethers.BrowserProvider(supportedProvider as unknown as ethers.Eip1193Provider);
+
+        // Initialize contract factories
+        this.weth9Factory = WETH9__factory;
+        this.exchangeProxyFactory = IZeroEx__factory;
     }
+
+    /**
+     * Get a WETH9 contract instance
+     */
+    public getWETH9Contract(address: string) {
+        return this.weth9Factory.connect(address, this._ethersProvider);
+    }
+
+    /**
+     * Get an IZeroEx contract instance
+     */
+    public getExchangeProxyContract(address: string) {
+        return this.exchangeProxyFactory.connect(address, this._ethersProvider);
+    }
+
     /**
      * Unsubscribes from all subscriptions for all contracts.
      */
     public unsubscribeAll(): void {
-        this.weth9.unsubscribeAll();
+        // TypeChain contracts don't require explicit unsubscription management
+        // Event listeners are managed per contract instance
     }
+
     /**
-     * Get the provider instance currently used by contract-wrappers
-     * @return  Web3 provider instance
+     * Get the provider instance
      */
-    public getProvider(): SupportedProvider {
-        return this._web3Wrapper.getProvider();
+    public getProvider() {
+        return this._supportedProvider;
     }
+
     /**
-     * Get the abi decoder instance currently used by contract-wrappers
-     * @return  AbiDecoder instance
+     * Get the ethers provider instance
      */
-    public getAbiDecoder(): AbiDecoder {
-        return this._web3Wrapper.abiDecoder;
+    public getEthersProvider() {
+        return this._ethersProvider;
     }
 }
