@@ -3,21 +3,17 @@ export {
     ContractEvent,
     ContractFunctionObj,
     ContractTxFunctionObj,
-    EncoderOverrides,
     SendTransactionOpts,
     SubscriptionErrors,
 } from '@0x/base-contract';
 export { ContractAddresses } from '@0x/contract-addresses';
 export {
-    DecodedLogEvent,
-    EventCallback,
-    IndexedFilterValues,
     SimpleContractArtifact,
     SimpleEvmBytecodeOutput,
     SimpleEvmOutput,
     SimpleStandardContractOutput,
-} from '@0x/types';
-export { AbiDecoder, AbiEncoder, DecodedCalldata, EncodingRules } from '@0x/utils';
+} from '@0x/utils';
+export { AbiDecoder, DecodedCalldata } from '@0x/utils';
 export {
     AbiDefinition,
     BlockParam,
@@ -48,25 +44,17 @@ export {
     EvmOutput,
     FallbackAbi,
     FunctionAbi,
-    GanacheProvider,
-    GethCallOverrides,
-    JSONRPCErrorCallback,
     JSONRPCRequestPayload,
     JSONRPCResponseError,
     JSONRPCResponsePayload,
     LogEntry,
-    LogEntryEvent,
     LogWithDecodedArgs,
     MethodAbi,
-    OptimizerSettings,
-    OutputField,
     ParamDescription,
-    RawLog,
-    RevertErrorAbi,
-    StandardContractOutput,
     StateMutability,
     SupportedProvider,
-    TupleDataItem,
+    TraceParams,
+    TransactionReceiptWithDecodedLogs,
     TxData,
     TxDataPayable,
     Web3JsProvider,
@@ -75,118 +63,29 @@ export {
     Web3JsV3Provider,
     ZeroExProvider,
 } from 'ethereum-types';
+
+// TypeChain exports with native bigint support
+export * from './typechain-types';
+
+// Re-export commonly used TypeChain types for easier access
+export type { IERC20Token, IEtherToken, WETH9, IZeroEx, ZRXToken } from './typechain-types';
+
+export type {
+    IERC20Token__factory,
+    IEtherToken__factory,
+    WETH9__factory,
+    IZeroEx__factory,
+    ZRXToken__factory,
+} from './typechain-types/factories';
+
+// Legacy types and configuration
+export {
+    ContractError,
+    ContractWrappersConfig,
+    ForwarderError,
+    OrderInfo,
+    OrderStatus,
+    TraderInfo,
+    OrderAndTraderInfo,
+} from './types';
 export { ContractWrappers } from './contract_wrappers';
-export { CoordinatorContract } from './generated-wrappers/coordinator';
-export { DevUtilsContract } from './generated-wrappers/dev_utils';
-export {
-    ERC20TokenApprovalEventArgs,
-    ERC20TokenContract,
-    ERC20TokenEventArgs,
-    ERC20TokenEvents,
-    ERC20TokenTransferEventArgs,
-} from './generated-wrappers/erc20_token';
-export {
-    ERC721TokenApprovalEventArgs,
-    ERC721TokenApprovalForAllEventArgs,
-    ERC721TokenContract,
-    ERC721TokenEventArgs,
-    ERC721TokenEvents,
-    ERC721TokenTransferEventArgs,
-} from './generated-wrappers/erc721_token';
-export {
-    ExchangeAssetProxyRegisteredEventArgs,
-    ExchangeCancelEventArgs,
-    ExchangeCancelUpToEventArgs,
-    ExchangeContract,
-    ExchangeEventArgs,
-    ExchangeEvents,
-    ExchangeFillEventArgs,
-    ExchangeOwnershipTransferredEventArgs,
-    ExchangeProtocolFeeCollectorAddressEventArgs,
-    ExchangeProtocolFeeMultiplierEventArgs,
-    ExchangeSignatureValidatorApprovalEventArgs,
-    ExchangeTransactionExecutionEventArgs,
-} from './generated-wrappers/exchange';
-export {
-    ForwarderContract,
-    ForwarderEventArgs,
-    ForwarderEvents,
-    ForwarderOwnershipTransferredEventArgs,
-} from './generated-wrappers/forwarder';
-export { IAssetDataContract } from './generated-wrappers/i_asset_data'; // used for synchronously encoding and decoding asset data
-export {
-    ITransformERC20Contract,
-    ITransformERC20EventArgs,
-    ITransformERC20Events,
-    ITransformERC20QuoteSignerUpdatedEventArgs,
-    ITransformERC20TransformedERC20EventArgs,
-    ITransformERC20TransformerDeployerUpdatedEventArgs,
-} from './generated-wrappers/i_transform_erc20';
-export {
-    IZeroExContract,
-    IZeroExEventArgs,
-    IZeroExEvents,
-    IZeroExLiquidityProviderSwapEventArgs,
-    IZeroExMetaTransactionExecutedEventArgs,
-    IZeroExMigratedEventArgs,
-    IZeroExOwnershipTransferredEventArgs,
-    IZeroExProxyFunctionUpdatedEventArgs,
-    IZeroExQuoteSignerUpdatedEventArgs,
-    IZeroExTransformedERC20EventArgs,
-    IZeroExTransformerDeployerUpdatedEventArgs,
-    IZeroExLimitOrderFilledEventArgs,
-    IZeroExOrderCancelledEventArgs,
-    IZeroExOtcOrderFilledEventArgs,
-    IZeroExPairCancelledLimitOrdersEventArgs,
-    IZeroExPairCancelledRfqOrdersEventArgs,
-    IZeroExRfqOrderFilledEventArgs,
-    IZeroExRfqOrderOriginsAllowedEventArgs,
-    IZeroExOrderSignerRegisteredEventArgs,
-    IZeroExERC1155OrderCancelledEventArgs,
-    IZeroExERC1155OrderFilledEventArgs,
-    IZeroExERC1155OrderPreSignedEventArgs,
-    IZeroExERC721OrderCancelledEventArgs,
-    IZeroExERC721OrderFilledEventArgs,
-    IZeroExERC721OrderPreSignedEventArgs,
-} from './generated-wrappers/i_zero_ex';
-export {
-    StakingAuthorizedAddressAddedEventArgs,
-    StakingAuthorizedAddressRemovedEventArgs,
-    StakingContract,
-    StakingEpochEndedEventArgs,
-    StakingEpochFinalizedEventArgs,
-    StakingEventArgs,
-    StakingEvents,
-    StakingExchangeAddedEventArgs,
-    StakingExchangeRemovedEventArgs,
-    StakingMakerStakingPoolSetEventArgs,
-    StakingMoveStakeEventArgs,
-    StakingOperatorShareDecreasedEventArgs,
-    StakingOwnershipTransferredEventArgs,
-    StakingParamsSetEventArgs,
-    StakingRewardsPaidEventArgs,
-    StakingStakeEventArgs,
-    StakingStakingPoolCreatedEventArgs,
-    StakingStakingPoolEarnedRewardsInEpochEventArgs,
-    StakingUnstakeEventArgs,
-} from './generated-wrappers/staking';
-export {
-    StakingProxyAuthorizedAddressAddedEventArgs,
-    StakingProxyAuthorizedAddressRemovedEventArgs,
-    StakingProxyContract,
-    StakingProxyEventArgs,
-    StakingProxyEvents,
-    StakingProxyOwnershipTransferredEventArgs,
-    StakingProxyStakingContractAttachedToProxyEventArgs,
-    StakingProxyStakingContractDetachedFromProxyEventArgs,
-} from './generated-wrappers/staking_proxy';
-export {
-    WETH9ApprovalEventArgs,
-    WETH9Contract,
-    WETH9DepositEventArgs,
-    WETH9EventArgs,
-    WETH9Events,
-    WETH9TransferEventArgs,
-    WETH9WithdrawalEventArgs,
-} from './generated-wrappers/weth9';
-export { ContractError, ContractWrappersConfig, ForwarderError, OrderInfo, OrderStatus } from './types';
