@@ -1,4 +1,2 @@
-import { BigNumber } from '@0x/utils';
-
 export const ETH_TOKEN_ADDRESS = '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee';
-export const ZERO = new BigNumber(0);
+export const ZERO = 0n;
